refactor(desktop): split ApiClient into per-domain facets over one HTTP core
api/client.ts was a 3,588-line class that conflicted on nearly every PR.
The bearer, JSON, byte, and delivery-timeout transport now lives in
api/client/http.ts as HttpCore, and each server route family adds its
methods through a mixin in its own module under api/client/. client.ts
composes the facets back into the same ApiClient class, so every call
site and its public exports compile unchanged.
