+++ crates/tidebreak-desktop/ui/src/api/client.ts
@@ -1,3558 +1,61 @@
-import { invoke, isTauri } from "@tauri-apps/api/core";
-
-import { attachedRemotely, attachTrustedFolders } from "../host";
-
-import {
-  APP_INVOKE_REFUSAL_KINDS,
-  AppInvokeRefusalError,
-  type AppDetail,
-  type AppFolderInvokeResult,
-  type AppGatewayPageResult,
-  type AppGrantState,
-  type AppInvokeRefusalKind,
-  type AppLibrary,
-  type AppRestInvokeResult,
-  type AppViewSession,
-  type ApprovalGrantRung,
-  type Chat,
-  type ChatFrame,
-  type ChatTranscript,
-  type ExecConfigInfo,
-  type ExecCredentialReadiness,
-  type ExecProviderKind,
-  type CompactionRun,
-  type ConnectedAppsInfo,
-  type ConsentStatementSnapshot,
-  type CustomModelConfig,
-  type DocumentDetail,
-  type EgressConfig,
-  type ExecFileUndoOutcome,
-  type FileDownloadProgress,
-  type GatewayApps,
-  type GatewayStatus,
-  type ManagedPolicy,
-  type McpAppPayload,
-  type McpServerDefinition,
-  type McpServersInfo,
-  type McpViewSession,
-  type ModelCatalog,
-  type ModelRole,
-  type ModelRoleInfo,
-  type ModelSelectionKey,
-  type ModelVisibility,
-  type NetworkPolicy,
-  type PendingFolderAccessRequest,
-  type PromptCacheRetention,
-  type QueuedCodeTurn,
-  type QueuedTurn,
-  type PendingOutputWritebackRequest,
-  type PendingPlanApproval,
-  type PendingToolApproval,
-  type PendingUserQuestions,
-  type PermissionMode,
-  type PlanDecision,
-  type PluginCatalog,
-  type PluginEnableUpdate,
-  type Project,
-  type ProjectDocumentPage,
-  type PromptBody,
-  type ProviderInfo,
-  type ProviderKind,
-  type ChatGptSignInStatus,
-  type ReasoningEffort,
-  type RestCredentialUpdate,
-  type RuntimeSettings,
-  type SandboxAgentCancellation,
-  type SkillInstructions,
-  type SpecPreviewInfo,
-  type StandingGrantSnapshot,
-  type UserQuestionAnswer,
-  type VoiceTranscriptionInfo,
-  type VoiceTranscriptionModel,
-  type WebSearchConfigInfo,
-  type WebSearchCredentialReadiness,
-  type WebSearchMode,
-  type WebSearchProviderKind,
-  type AgentActivityHistoryEntry,
-  type AgentRun,
-  type AgentRunProgress,
-  type LocalVoiceInfo,
-  type InboxEntry,
-  type AgentNotificationPage,
-  inboxConversationKey,
-  type AgentRunTaskPlan,
-  type PendingChatPrompt,
-  type TaskPlan,
-  type CodeApprovalSnapshot,
-  type CodeRepoSnapshot,
-  type QuickAction,
-  type CodeSessionSnapshot,
-  type CodeGrantSnapshot,
-  type CodeConnectPage,
-  type CodeTurnSnapshot,
-  type CodeActionSnapshot,
-  type CodeCommitSnapshot,
-  type CodePushSnapshot,
-  type CodeTerminalRead,
-  type CodeTerminalSnapshot,
-  type CodeWorkspaceDiff,
-  type CodeWorkspaceFiles,
-  type CodeWorkspaceSearch,
-  type CodeWorkspaceBlob,
-  type CodeWorkspaceTree,
-  type CodeWorkspacePrSnapshot,
-  type CodeWorkspacePullRequests,
-  type PullRequestDigest,
-  type CodeTriggerAction,
-  type CodeTriggerCondition,
-  type CodeTriggerSnapshot,
-  type CodeWatchSnapshot,
-  type CodePrCommentsSnapshot,
-  type CodePrMergeMethod,
-  type CodeWorkspaceSnapshot,
-  type CodeCloneDefaults,
-  type CodeGithubRepositories,
-  type CodeRepoSources,
-  type CodeWorktreeRoot,
-  type CodeCheckLogsSnapshot,
-  type CodeForkTranscript,
-  type CodeCloneJobSnapshot,
-  type CodeHarnessInstallSnapshot,
-  type CodeAnalyticsRange,
-  type CodeAnalyticsSnapshot,
-  type CodeSubscriptionUsage,
-  type CodeDeliveryActionResult,
-  type CodeDeliveryPullRequestActionBody,
-  type CodeDeliveryPullRequestDetail,
-  type CodeDeliveryPullRequestQuery,
-  type CodeDeliveryPullRequestTarget,
-  type CodeDeliveryPullRequestsPage,
-  type CodeDeliveryRepositoriesSnapshot,
-  type CodeDeliveryRunActionBody,
-  type CodeDeliveryRunDetail,
-  type CodeDeliveryRunQuery,
-  type CodeDeliveryRunTarget,
-  type CodeDeliveryRunsPage,
-  type HarnessDoctorReport,
-  type HarnessKind,
-  type SequencedCodeEventFrame,
-  type CodeUpdateNotice,
-} from "./types";
-import {
-  parseAgentActivityHistory,
-  parseAgentRunProgress,
-  parseFolderAccessRequest,
-  parseInboxEntry,
-  parseAgentNotificationPage,
-  parseOutputWritebackRequest,
-  parsePendingChatPrompt,
-  parsePendingPlanApproval,
-  parsePendingToolApproval,
-  parsePendingUserQuestions,
-  parseAgentRunTaskPlan,
-  parseSandboxAgentCancellation,
-  parseTaskPlan,
-} from "./parsers";
-import {
-  parseCodeApproval,
-  parseCodeCloneDefaults,
-  parseCodeGithubRepositories,
-  parseCodeRepoSources,
-  parseCodeWorktreeRoot,
-  parseCodeCheckLogsSnapshot,
-  parseCodeForkTranscript,
-  parseCodeCloneJob,
-  parseCodeHarnessInstall,
-  parseCodeAnalytics,
-  parseCodeRepo,
-  parseCodeSession,
-  parseCodeSessionList,
-  parseCodeGrantList,
-  parseCodeGrant,
-  parseCodeConnectPage,
-  parseCodeTurnList,
-  parseCodeTurnSubmission,
-  type CodeTurnSubmission,
-  parseQueuedCodeTurn,
-  parseCodeAction,
-  parseCodeCommit,
-  parseCodePush,
-  parseCodeTerminal,
-  parseCodeTerminalList,
-  parseCodeTerminalRead,
-  parseCodeWorkspace,
-  parseCodeWorkspaceDiff,
-  parseCodeWorkspaceFiles,
-  parseCodeWorkspaceSearch,
-  parseCodeWorkspaceBlob,
-  parseCodeWorkspaceTree,
-  parseCodeWorkspacePr,
-  parseCodeWorkspacePullRequests,
-  parseCodeTrigger,
-  parseCodeTriggers,
-  parseCodeWatch,
-  parseCodePrComments,
-  parseHarnessModelList,
-  type ParsedHarnessModelList,
-  parseHarnessDoctorReport,
-  parseSequencedCodeEvent,
-  parseCodeUpdateNotice,
-  parseCodeSubscriptionUsage,
-  parseCodeDeliveryActionResult,
-  parseCodeDeliveryPullRequestDetail,
-  parseCodeDeliveryPullRequestsPage,
-  parseCodeDeliveryRepositories,
-  parseCodeDeliveryRunDetail,
-  parseCodeDeliveryRunsPage,
-} from "../code/parsers";
-
-function parseMarked(body: unknown): number {
-  if (
-    !body ||
-    typeof body !== "object" ||
-    !("marked" in body) ||
-    typeof (body as { marked: unknown }).marked !== "number"
-  ) {
-    throw new Error("notification mark response is invalid");
-  }
-  return (body as { marked: number }).marked;
-}
-
-const WS_HANDSHAKE = "tidebreak-v1";
-const WS_TOKEN_PREFIX = "tidebreak-token.";
-const DEFAULT_DELIVERY_TIMEOUT_MS = 30_000;
-
-export type DeliveryRequestOptions = {
-  signal?: AbortSignal;
-  timeoutMs?: number;
-  refreshAuth?: boolean;
-};
-
-export type CodeWorkspaceMergeRequest = {
-  target: CodeDeliveryPullRequestTarget;
-  expected_head_sha: string;
-  method: CodePrMergeMethod;
-  auto?: boolean;
-};
-
-type CodeWorkspaceMergeResult = {
-  target: CodeDeliveryPullRequestTarget;
-  accepted_head_sha: string;
-  status: CodeWorkspacePrSnapshot;
-};
+import { withSettingsApi } from "./client/settings";
+import { withAppsApi } from "./client/apps";
+import { withProjectsApi } from "./client/projects";
+import { withChatApi } from "./client/chat";
+import { withAgentRunsApi } from "./client/agent-runs";
+import { withTurnsApi } from "./client/turns";
+import { withDeliveryApi } from "./client/delivery";
+import { withCodeReposApi } from "./client/code-repos";
+import { withCodeWorkspacesApi } from "./client/code-workspaces";
+import { withCodeSessionsApi } from "./client/code-sessions";
+import { withCodeFilesApi } from "./client/code-files";
+import { withCodeGitApi } from "./client/code-git";
+import { withCodeGrantsApi } from "./client/code-grants";
+import { withCodeTerminalsApi } from "./client/code-terminals";
+import { withCodeEventsApi } from "./client/code-events";
+import { HttpCore } from "./client/http";
+export {
+  ARCHIVE_FORCE_KINDS,
+  HttpError,
+  archiveForceKind,
+  type DeliveryRequestOptions,
+} from "./client/http";
+export { type CodeWorkspaceMergeRequest } from "./client/code-git";
 
 /**
- * The size below which a transfer is not worth reporting on.
+ * The desktop's HTTP and WebSocket client, assembled from per-domain facets
+ * over one HTTP core.
  *
- * A bar that appears and vanishes is worse than no bar, and a source under this
- * arrives in a couple of chunks — most of them from a sidecar on this machine.
- * Well under the 16 MB a source may be, so the files big enough to wait on do
- * still report.
+ * Each module under `./client/` owns one server route family and adds its
+ * methods through a mixin. This class only composes them, so call sites keep
+ * `client.method()` and a change to one route family touches one file.
  */
-const PROGRESS_MIN_BYTES = 2 * 1024 * 1024;
-
-/** Progress updates are worth at most one re-render per frame budget. */
-const PROGRESS_THROTTLE_MS = 100;
-
-/**
- * Rate-limit progress callbacks, always letting the last one through.
- *
- * Without the trailing call the bar can stop short of the end: the final chunk
- * usually lands inside the throttle window of the one before it.
- */
-function throttle(
-  report: (progress: FileDownloadProgress) => void,
-): (progress: FileDownloadProgress) => void {
-  let last = 0;
-  return (progress) => {
-    const now = Date.now();
-    if (
-      progress.loaded >= progress.total ||
-      now - last >= PROGRESS_THROTTLE_MS
-    ) {
-      last = now;
-      report(progress);
-    }
-  };
-}
-
-/**
- * A rejected response, carrying the status so a caller can tell why.
- *
- * The status is what separates "this is gone" from "we could not reach the
- * server": a panel that cannot tell those apart has to guess, and guessing
- * wrong tells a reader their file was deleted when it was not.
- */
-export class HttpError extends Error {
-  constructor(
-    readonly status: number,
-    message: string,
-    readonly kind?: string,
-  ) {
-    super(message);
-    this.name = "HttpError";
-  }
-}
-
-/** Archive is refused until the reader confirms discarding leftover work. */
-export const ARCHIVE_FORCE_KINDS = new Set([
-  "uncommitted",
-  "unpushed",
-  "uncommitted_and_unpushed",
-  "ignored_content",
-]);
-
-/** The 409 kinds that mean archive needs an explicit force. */
-export function archiveForceKind(error: unknown): string | null {
-  if (!(error instanceof HttpError) || !error.kind) return null;
-  return ARCHIVE_FORCE_KINDS.has(error.kind) ? error.kind : null;
-}
-
-/** The server's own message for a failed response, or its status text. */
-async function throwIfNotOk(response: Response): Promise<void> {
-  if (response.ok) return;
-  let detail = response.statusText;
-  let kind: string | undefined;
-  try {
-    const body = (await response.json()) as { message?: string; kind?: string };
-    if (body.message) detail = body.message;
-    if (typeof body.kind === "string" && body.kind.length > 0) kind = body.kind;
-  } catch {
-    /* ignore */
-  }
-  throw new HttpError(response.status, `${response.status}: ${detail}`, kind);
-}
-
-export class ApiClient {
-  private accessToken: string;
-  private readonly workspaceMergeTargets = new Map<
-    string,
-    Pick<CodeWorkspaceMergeRequest, "target" | "expected_head_sha">
-  >();
-
-  constructor(
-    readonly baseUrl: string,
-    token: string,
-  ) {
-    this.accessToken = token;
-  }
-
-  get token(): string {
-    return this.accessToken;
-  }
-
-  /** Replace the short-lived bearer used by subsequent HTTP and WebSocket connections. */
-  setAccessToken(token: string): void {
-    this.accessToken = token;
-  }
-
-  private headers(json = false): HeadersInit {
-    const headers: Record<string, string> = {
-      Authorization: `Bearer ${this.token}`,
-    };
-    if (json) headers["Content-Type"] = "application/json";
-    return headers;
-  }
-
-  private async json<T>(
-    path: string,
-    init?: RequestInit,
-    expectedStatus?: number,
-  ): Promise<T> {
-    const response = await fetch(`${this.baseUrl}${path}`, init);
-    await throwIfNotOk(response);
-    if (expectedStatus !== undefined && response.status !== expectedStatus) {
-      throw new Error(
-        `unexpected response status: expected ${expectedStatus}, received ${response.status}`,
-      );
-    }
-    if (response.status === 204) return undefined as T;
-    const text = await response.text();
-    if (text.length === 0) return undefined as T;
-    return JSON.parse(text) as T;
-  }
-
-  private async deliveryJson<T>(
-    path: string,
-    init: RequestInit,
-    options: DeliveryRequestOptions = {},
-  ): Promise<T> {
-    const controller = new AbortController();
-    const timeoutMs = options.timeoutMs ?? DEFAULT_DELIVERY_TIMEOUT_MS;
-    let timedOut = false;
-    const abortFromCaller = () => controller.abort(options.signal?.reason);
-    if (options.signal?.aborted) abortFromCaller();
-    else
-      options.signal?.addEventListener("abort", abortFromCaller, {
-        once: true,
-      });
-    const timeout = globalThis.setTimeout(() => {
-      timedOut = true;
-      controller.abort();
-    }, timeoutMs);
-    try {
-      return await this.json<T>(path, { ...init, signal: controller.signal });
-    } catch (error) {
-      if (timedOut) {
-        throw new Error(
-          `GitHub delivery request timed out after ${Math.ceil(timeoutMs / 1_000)} seconds.`,
-        );
-      }
-      throw error;
-    } finally {
-      globalThis.clearTimeout(timeout);
-      options.signal?.removeEventListener("abort", abortFromCaller);
-    }
-  }
-  listProviders(): Promise<{ providers: ProviderInfo[] }> {
-    return this.json("/providers", { headers: this.headers() });
-  }
-
-  putProvider(
-    kind: ProviderKind,
-    body: {
-      enabled?: boolean;
-      base_url?: string | null;
-      credential?: { type: "api_key"; key: string };
-      models?: CustomModelConfig[];
-    },
-  ): Promise<ProviderInfo> {
-    return this.json(`/providers/${kind}`, {
-      method: "PUT",
-      headers: this.headers(true),
-      body: JSON.stringify(body),
-    });
-  }
-
-  deleteCredential(kind: ProviderKind): Promise<void> {
-    return this.json(`/providers/${kind}/credential`, {
-      method: "DELETE",
-      headers: this.headers(),
-    });
-  }
-
-  openaiChatgptSignIn(): Promise<{ authorization_url: string }> {
-    return this.json("/providers/openai/chatgpt/sign-in", {
-      method: "POST",
-      headers: this.headers(),
-    });
-  }
-
-  openaiChatgptSignOut(): Promise<void> {
-    return this.json("/providers/openai/chatgpt/sign-out", {
-      method: "POST",
-      headers: this.headers(),
-    });
-  }
-
-  getOpenaiChatgptStatus(): Promise<ChatGptSignInStatus> {
-    return this.json("/providers/openai/chatgpt/status", {
-      headers: this.headers(),
-    });
-  }
-
-  getVoiceTranscription(): Promise<VoiceTranscriptionInfo> {
-    return this.json("/voice-transcription", { headers: this.headers() });
-  }
-
-  putVoiceTranscription(
-    model: VoiceTranscriptionModel,
-    localModel?: string,
-  ): Promise<VoiceTranscriptionInfo> {
-    return this.json("/voice-transcription", {
-      method: "PUT",
-      headers: this.headers(true),
-      body: JSON.stringify({ model, local_model: localModel ?? null }),
-    });
-  }
-
-  installLocalVoice(model: string): Promise<LocalVoiceInfo> {
-    return this.json("/voice-transcription/install", {
-      method: "POST",
-      headers: this.headers(true),
-      body: JSON.stringify({ model }),
-    });
-  }
-
-  async transcribeVoice(audio: Blob): Promise<string> {
-    const response = await fetch(`${this.baseUrl}/voice-transcription`, {
-      method: "POST",
-      headers: {
-        ...this.headers(),
-        "Content-Type": audio.type || "audio/webm",
-      },
-      body: audio,
-    });
-    await throwIfNotOk(response);
-    return ((await response.json()) as { text: string }).text;
-  }
-
-  /**
-   * The selectable catalog, plus one row per model role: what the user pinned
-   * it to, and what it resolves to right now — the only way a client can name
-   * what "default" or "automatic" means for a role.
-   */
-  listModels(): Promise<ModelCatalog> {
-    return this.json("/models", { headers: this.headers() });
-  }
-
-  /**
-   * Pin a model role to one model, or pass `null` to return it to automatic
-   * resolution against the role's ordered defaults.
-   */
-  putModelRole(
-    role: ModelRole,
-    selection: ModelSelectionKey | null,
-  ): Promise<ModelRoleInfo> {
-    return this.json(`/models/roles/${role}`, {
-      method: "PUT",
-      headers: this.headers(true),
-      body: JSON.stringify({ selection }),
-    });
-  }
-
-  getSettings(): Promise<RuntimeSettings> {
-    return this.json("/settings", { headers: this.headers() });
-  }
-
-  /**
-   * Update runtime settings. A field absent leaves it unchanged, `null` resets
-   * it to the server default, and a value sets it (matching the double-option
-   * body the server expects).
-   *
-   * `model_visibility_overrides` is the exception to "absent leaves it
-   * unchanged, present merges": the server replaces the map wholesale, so a
-   * writer sends the complete set of deviations it wants persisted.
-   */
-  putSettings(body: {
-    model?: ModelSelectionKey | null;
-    max_active_background_agents?: number;
-    sandbox_agent_checkin_steps?: number;
-    sandbox_agent_error_checkin?: number;
-    model_visibility_overrides?: Record<string, ModelVisibility>;
-    prompt_cache_retention?: PromptCacheRetention;
-    compaction?: {
-      threshold_fraction?: number;
-      target_fraction?: number;
-      min_threshold_tokens?: number;
-      protect_recent_messages?: number;
-    };
-    computer_use_enabled?: boolean;
-    code_turn_recaps_enabled?: boolean;
-    rewrite_closing_messages?: boolean;
-    git_source_control?: {
-      auto_rename_branches?: boolean;
-      branch_prefix_mode?: "account" | "custom" | "none";
-      custom_branch_prefix?: string | null;
-    };
-  }): Promise<RuntimeSettings> {
-    return this.json("/settings", {
-      method: "PUT",
-      headers: this.headers(true),
-      body: JSON.stringify(body),
-    });
-  }
-
-  getWebSearchConfig(): Promise<WebSearchConfigInfo> {
-    return this.json("/web-search", { headers: this.headers() });
-  }
-
-  putWebSearchConfig(body: {
-    mode?: WebSearchMode;
-    provider?: WebSearchProviderKind | null;
-    timeout_ms?: number;
-    // Explicit null clears the configured instance URL; omitting the field
-    // leaves it as it is.
-    searxng_base_url?: string | null;
-  }): Promise<WebSearchConfigInfo> {
-    return this.json("/web-search", {
-      method: "PUT",
-      headers: this.headers(true),
-      body: JSON.stringify(body),
-    });
-  }
-
-  listWebSearchCredentials(): Promise<{
-    credentials: WebSearchCredentialReadiness[];
-  }> {
-    return this.json("/web-search/credentials", { headers: this.headers() });
-  }
-
-  putWebSearchCredential(
-    provider: WebSearchProviderKind,
-    apiKey: string,
-  ): Promise<WebSearchCredentialReadiness> {
-    return this.json(`/web-search/credentials/${provider}`, {
-      method: "PUT",
-      headers: this.headers(true),
-      body: JSON.stringify({ api_key: apiKey }),
-    });
-  }
-
-  deleteWebSearchCredential(
-    provider: WebSearchProviderKind,
-  ): Promise<WebSearchCredentialReadiness> {
-    return this.json(`/web-search/credentials/${provider}`, {
-      method: "DELETE",
-      headers: this.headers(),
-    });
-  }
-
-  getExecConfig(): Promise<ExecConfigInfo> {
-    return this.json("/code-execution", { headers: this.headers() });
-  }
-
-  putExecConfig(body: {
-    provider?: ExecProviderKind | null;
-    timeout_ms?: number;
-    egress?: EgressConfig;
-  }): Promise<ExecConfigInfo> {
-    return this.json("/code-execution", {
-      method: "PUT",
-      headers: this.headers(true),
-      body: JSON.stringify(body),
-    });
-  }
-
-  listExecCredentials(): Promise<{
-    credentials: ExecCredentialReadiness[];
-  }> {
-    return this.json("/code-execution/credentials", {
-      headers: this.headers(),
-    });
-  }
-
-  putExecCredential(
-    provider: ExecProviderKind,
-    apiKey: string,
-  ): Promise<ExecCredentialReadiness> {
-    return this.json(`/code-execution/credentials/${provider}`, {
-      method: "PUT",
-      headers: this.headers(true),
-      body: JSON.stringify({ api_key: apiKey }),
-    });
-  }
-
-  deleteExecCredential(
-    provider: ExecProviderKind,
-  ): Promise<ExecCredentialReadiness> {
-    return this.json(`/code-execution/credentials/${provider}`, {
-      method: "DELETE",
-      headers: this.headers(),
-    });
-  }
-
-  listConnectedApps(): Promise<ConnectedAppsInfo> {
-    return this.json("/connected-apps", { headers: this.headers() });
-  }
-
-  putRestConnectedApp(
-    id: string,
-    body: {
-      name: string;
-      base_url: string;
-      /** The raw JSON OpenAPI document, when supplied inline; the server
-       * ingests it once here. Exactly one of this and
-       * `openapi_document_url` must be present. */
-      openapi_document?: string;
-      /** URL the server fetches the document from at save time. Requires
-       * `document_sha256`. */
-      openapi_document_url?: string;
-      /** The preview's document hash pin; the save refuses (409) if the
-       * document no longer matches it. */
-      document_sha256?: string;
-      /** When present, only these operationIds are ingested; the rest of
-       * the document is not judged. */
-      operation_ids?: string[];
-      credential: RestCredentialUpdate;
-    },
-  ): Promise<ConnectedAppsInfo> {
-    return this.json(`/connected-apps/rest/${encodeURIComponent(id)}`, {
-      method: "PUT",
-      headers: this.headers(true),
-      body: JSON.stringify(body),
-    });
-  }
-
-  /** List what an OpenAPI document declares, for the operation picker. */
-  previewRestSpec(
-    source: { url: string } | { document: string },
-  ): Promise<SpecPreviewInfo> {
-    return this.json("/connected-apps/rest/spec-preview", {
-      method: "POST",
-      headers: this.headers(true),
-      body: JSON.stringify({ source }),
-    });
-  }
-
-  deleteRestConnectedApp(id: string): Promise<void> {
-    return this.json(`/connected-apps/rest/${encodeURIComponent(id)}`, {
-      method: "DELETE",
-      headers: this.headers(),
-    });
-  }
-
-  listMcpServers(): Promise<McpServersInfo> {
-    return this.json("/mcp/servers", { headers: this.headers() });
-  }
-
-  putMcpServers(servers: McpServerDefinition[]): Promise<McpServersInfo> {
-    // The native command exists to put an OS confirmation in front of stdio
-    // commands that would run on this computer, and it writes to the server
-    // embedded in this app. Neither is right while attached: the commands
-    // would run on the machine, and writing here would save the reader's
-    // edit to a config they were not looking at — `listMcpServers` reads the
-    // machine's, so the panel would re-read and show the edit gone.
-    if (isTauri() && !attachedRemotely()) {
-      return invoke<McpServersInfo>("put_native_mcp_servers", {
-        config: { servers },
-      });
-    }
-    return this.json("/mcp/servers", {
-      method: "PUT",
-      headers: this.headers(true),
-      body: JSON.stringify({ servers }),
-    });
-  }
-
-  reconnectMcpServer(name: string): Promise<McpServersInfo> {
-    return this.json(`/mcp/servers/${encodeURIComponent(name)}/reconnect`, {
-      method: "POST",
-      headers: this.headers(),
-    });
-  }
-
-  /** Trade the bearer for a single-use iframe address for one view. */
-  createMcpViewFrame(server: string, uri: string): Promise<McpViewSession> {
-    return this.json(
-      `/mcp/servers/${encodeURIComponent(server)}/view-session`,
-      {
-        method: "POST",
-        headers: this.headers(true),
-        body: JSON.stringify({ uri }),
-      },
-    );
-  }
-
-  getPolicy(): Promise<ManagedPolicy> {
-    return this.json("/policy", { headers: this.headers() });
-  }
-
-  getGatewayStatus(): Promise<GatewayStatus> {
-    return this.json("/gateway/status", { headers: this.headers() });
-  }
-
-  gatewaySignIn(): Promise<{ authorization_url: string }> {
-    return this.json("/gateway/sign-in", {
-      method: "POST",
-      headers: this.headers(),
-    });
-  }
-
-  gatewaySignOut(): Promise<GatewayStatus> {
-    return this.json("/gateway/sign-out", {
-      method: "POST",
-      headers: this.headers(),
-    });
-  }
-
-  /** Decline the pending deep-link pairing; returns the policy to render. */
-  dismissGatewayPairing(): Promise<ManagedPolicy> {
-    return this.json("/gateway/pairing/dismiss", {
-      method: "POST",
-      headers: this.headers(),
-    });
-  }
-
-  getGatewayApps(): Promise<GatewayApps> {
-    return this.json("/gateway/apps", { headers: this.headers() });
-  }
-
-  /**
-   * The hosted machine this profile's gateway offers, if it offers one.
-   *
-   * A hint for the address field, never a grant: attaching runs the same
-   * discovery handshake either way. An absent `url` means no prefill — a
-   * gateway that hosts no machine, one older than the field, and one that
-   * did not answer are the same answer here.
-   */
-  getGatewayMachine(): Promise<{ url?: string }> {
-    return this.json("/gateway/machine", { headers: this.headers() });
-  }
-
-  syncGatewayModels(): Promise<GatewayStatus> {
-    return this.json("/gateway/models/sync", {
-      method: "POST",
-      headers: this.headers(),
-    });
-  }
-
-  getMcpAppPayload(chatId: string, callId: string): Promise<McpAppPayload> {
-    return this.json(
-      `/chats/${encodeURIComponent(chatId)}/calls/${encodeURIComponent(callId)}/mcp-app-payload`,
-      { headers: this.headers() },
-    );
-  }
-
-  listApps(): Promise<AppLibrary> {
-    return this.json("/apps", { headers: this.headers() });
-  }
-
-  listPlugins(): Promise<PluginCatalog> {
-    return this.json("/plugins", { headers: this.headers() });
-  }
-
-  /** One skill's full instruction body — what the model is taught by it. */
-  getSkillInstructions(name: string): Promise<SkillInstructions> {
-    return this.json(
-      `/plugins/skills/${encodeURIComponent(name)}/instructions`,
-      {
-        headers: this.headers(),
-      },
-    );
-  }
-
-  /**
-   * One prompt's insertable text — what a picker drops into the composer.
-   *
-   * Its own route because the catalog is read far more often than any single
-   * prompt is inserted.
-   */
-  getPromptBody(name: string): Promise<PromptBody> {
-    return this.json(`/plugins/prompts/${encodeURIComponent(name)}/body`, {
-      headers: this.headers(),
-    });
-  }
-
-  /**
-   * Set the named enable flags, and take the fresh catalog back.
-   *
-   * A merge patch: the body names only what is changing, and the response is
-   * the authority on what the whole catalog now looks like — which is what
-   * lets a surface toggle optimistically and reconcile from one round trip.
-   */
-  setPluginsEnabled(update: PluginEnableUpdate): Promise<PluginCatalog> {
-    return this.json("/plugins/enabled", {
-      method: "PUT",
-      headers: this.headers(true),
-      body: JSON.stringify(update),
-    });
-  }
-
-  getApp(appId: string): Promise<AppDetail> {
-    return this.json(`/apps/${encodeURIComponent(appId)}`, {
-      headers: this.headers(),
-    });
-  }
-
-  deleteApp(appId: string): Promise<void> {
-    return this.json(`/apps/${encodeURIComponent(appId)}`, {
-      method: "DELETE",
-      headers: this.headers(),
-    });
-  }
-
-  getAppGrant(appId: string): Promise<AppGrantState> {
-    return this.json(`/apps/${encodeURIComponent(appId)}/grant`, {
-      headers: this.headers(),
-    });
-  }
-
-  /**
-   * Record consent. Deliberately body-less: consent is only ever "yes to what
-   * the server shows right now" — the server recomputes the grant from the
-   * current manifest and definitions, so a stale sheet can never widen it.
-   */
-  consentAppGrant(appId: string): Promise<AppGrantState> {
-    return this.json(`/apps/${encodeURIComponent(appId)}/grant`, {
-      method: "POST",
-      headers: this.headers(),
-    });
-  }
-
-  revokeAppGrant(appId: string): Promise<void> {
-    return this.json(`/apps/${encodeURIComponent(appId)}/grant`, {
-      method: "DELETE",
-      headers: this.headers(),
-    });
-  }
-
-  /**
-   * Where this app's page lives at the gateway, registering it there first if
-   * it has never been registered.
-   *
-   * Publishing itself happens on that page, not here: it mutates entitlement
-   * state the gateway owns, and every other mutation of that state is already
-   * done at the gateway (decision record 14).
-   *
-   * Every way the gateway can decline is an outcome in the response rather
-   * than a thrown error: the author asked a legitimate question and the
-   * gateway's answer — including its own words — is what the page renders.
-   */
-  appGatewayPage(appId: string): Promise<AppGatewayPageResult> {
-    return this.json(`/apps/${encodeURIComponent(appId)}/gateway-page`, {
-      method: "POST",
-      headers: this.headers(true),
-    });
-  }
-
-  /** Trade the bearer for a single-use iframe address for one app revision. */
-  createAppViewFrame(appId: string): Promise<AppViewSession> {
-    return this.json(`/apps/${encodeURIComponent(appId)}/view-session`, {
-      method: "POST",
-      headers: this.headers(true),
-      body: JSON.stringify({}),
-    });
-  }
-
-  /**
-   * Execute one of an app's pinned REST operations outside any turn.
-   * `parameters`, `body`, and the result are opaque passthrough between the
-   * sandboxed frame and the server; a typed refusal surfaces as
-   * {@link AppInvokeRefusalError} so the caller can branch on
-   * `consent_required` without string-matching prose. The response body
-   * crosses base64-encoded in `body_base64` (see {@link AppRestInvokeResult}).
-   */
-  async invokeAppOperation(
-    appId: string,
-    operationId: string,
-    parameters?: unknown,
-    body?: unknown,
-  ): Promise<AppRestInvokeResult> {
-    const request: Record<string, unknown> = { operation_id: operationId };
-    if (parameters !== undefined) request.parameters = parameters;
-    if (body !== undefined) request.body = body;
-    return (await this.postAppInvoke(appId, request)) as AppRestInvokeResult;
-  }
-
-  /**
-   * Execute one operation of an app's granted gateway binding — the same
-   * invoke route, relayed by the server to the model gateway as the signed-in
-   * user. The field names are the gateway's own invoke vocabulary, so a
-   * bundle authored here speaks the same shape to the gateway's shell;
-   * `path_parameters`, `query`, `body`, and the result are opaque passthrough.
-   */
-  async invokeAppGatewayOperation(
-    appId: string,
-    gatewayApp: string,
-    operationId: string,
-    pathParameters?: unknown,
-    query?: unknown,
-    body?: unknown,
-  ): Promise<AppRestInvokeResult> {
-    const request: Record<string, unknown> = {
-      gateway_app: gatewayApp,
-      operation_id: operationId,
-    };
-    if (pathParameters !== undefined) request.path_parameters = pathParameters;
-    if (query !== undefined) request.query = query;
-    if (body !== undefined) request.body = body;
-    return (await this.postAppInvoke(appId, request)) as AppRestInvokeResult;
-  }
-
-  /**
-   * Execute one folder operation of an app's granted folder binding — the
-   * `folder` sibling of {@link invokeAppOperation}, with the same refusal
-   * contract. File content crosses base64-encoded in both directions;
-   * failures come back as `is_error` results in the host's closed
-   * vocabulary.
-   */
-  async invokeAppFolder(
-    appId: string,
-    folder: string,
-    op: "list" | "read" | "write",
-    path?: string,
-    contentBase64?: string,
-    replace?: boolean,
-  ): Promise<AppFolderInvokeResult> {
-    const request: Record<string, unknown> = { folder, op };
-    if (path !== undefined) request.path = path;
-    if (contentBase64 !== undefined) request.content_base64 = contentBase64;
-    if (replace !== undefined) request.replace = replace;
-    return (await this.postAppInvoke(appId, request)) as AppFolderInvokeResult;
-  }
-
-  /** The shared invoke POST: one route, typed refusals surfaced as errors. */
-  private async postAppInvoke(
-    appId: string,
-    request: unknown,
-  ): Promise<unknown> {
-    const response = await fetch(
-      `${this.baseUrl}/apps/${encodeURIComponent(appId)}/invoke`,
-      {
-        method: "POST",
-        headers: this.headers(true),
-        body: JSON.stringify(request),
-      },
-    );
-    if (!response.ok) {
-      let refusal: unknown;
-      try {
-        refusal = await response.clone().json();
-      } catch {
-        /* not a typed refusal; fall through to the generic error */
-      }
-      if (
-        typeof refusal === "object" &&
-        refusal !== null &&
-        "kind" in refusal &&
-        "message" in refusal &&
-        APP_INVOKE_REFUSAL_KINDS.includes(
-          (refusal as { kind: AppInvokeRefusalKind }).kind,
-        )
-      ) {
-        const typed = refusal as {
-          kind: AppInvokeRefusalKind;
-          message: string;
-        };
-        throw new AppInvokeRefusalError(typed.kind, String(typed.message));
-      }
-      await throwIfNotOk(response);
-    }
-    return await response.json();
-  }
-
-  createProject(title: string): Promise<Project> {
-    return this.json("/projects", {
-      method: "POST",
-      headers: this.headers(true),
-      body: JSON.stringify({ title }),
-    });
-  }
-
-  listProjects(): Promise<Project[]> {
-    return this.json("/projects", { headers: this.headers() });
-  }
-
-  patchProjectTitle(projectId: string, title: string | null): Promise<Project> {
-    return this.json(`/projects/${encodeURIComponent(projectId)}`, {
-      method: "PATCH",
-      headers: this.headers(true),
-      body: JSON.stringify({ title }),
-    });
-  }
-
-  deleteProject(projectId: string): Promise<void> {
-    return this.json(`/projects/${encodeURIComponent(projectId)}`, {
-      method: "DELETE",
-      headers: this.headers(),
-    });
-  }
-
-  /** The files this project shares with every conversation filed under it. */
-  listProjectDocuments(projectId: string): Promise<ProjectDocumentPage> {
-    return this.json(`/projects/${encodeURIComponent(projectId)}/documents`, {
-      headers: this.headers(),
-    });
-  }
-
-  /**
-   * Share one conversation's file with the project that conversation belongs to.
-   *
-   * The conversation keeps its own copy — a document's owner is part of its id,
-   * so the project's is a different document and the transcript that referred to
-   * the original still resolves.
-   */
-  promoteDocumentToProject(
-    projectId: string,
-    chatId: string,
-    documentId: string,
-  ): Promise<{ document_id: string }> {
-    return this.json(
-      `/projects/${encodeURIComponent(projectId)}/documents/promote`,
-      {
-        method: "POST",
-        headers: this.headers(true),
-        body: JSON.stringify({ chat_id: chatId, document_id: documentId }),
-      },
-    );
-  }
-
-  deleteProjectDocument(projectId: string, documentId: string): Promise<void> {
-    return this.json(
-      `/projects/${encodeURIComponent(projectId)}/documents/${encodeURIComponent(documentId)}`,
-      { method: "DELETE", headers: this.headers() },
-    );
-  }
-
-  /**
-   * Create a chat, optionally already set up the way it will run.
-   *
-   * The turn inputs are sent at creation rather than PATCHed afterwards: a
-   * correcting PATCH races the first turn, which reads the chat as it was
-   * created.
-   */
-  async createChat(
-    model?: ModelSelectionKey,
-    projectId?: string | null,
-    settings?: {
-      reasoningEffort?: ReasoningEffort | null;
-      permissionMode?: PermissionMode | null;
-      networkPolicy?: NetworkPolicy;
-    },
-  ): Promise<Chat> {
-    const chat = await this.json<Chat>("/chats", {
-      method: "POST",
-      headers: this.headers(true),
-      body: JSON.stringify({
-        model: model || undefined,
-        project_id: projectId || undefined,
-        reasoning_effort: settings?.reasoningEffort ?? undefined,
-        permission_mode: settings?.permissionMode ?? undefined,
-        network_policy: settings?.networkPolicy,
-      }),
-    });
-    try {
-      await attachTrustedFolders(chat);
-    } catch (error) {
-      // The chat already exists. Keep it usable if the local folder broker is
-      // unavailable instead of making a retry create a second conversation.
-      console.warn("could not attach saved folders to the new chat", error);
-    }
-    return chat;
-  }
-
-  listChats(): Promise<Chat[]> {
-    return this.json("/chats", { headers: this.headers() });
-  }
-
-  getChat(chatId: string): Promise<Chat> {
-    return this.json(`/chats/${chatId}`, { headers: this.headers() });
-  }
-
-  /**
-   * Everything parked on this reader, across their conversations.
-   *
-   * One server-side read rather than a loop over chats: the shell needs the
-   * whole set to badge the inbox and mark the rail, and asking each chat in
-   * turn would make that cost grow with the profile.
-   */
-  async listNotifications(cursor?: string): Promise<AgentNotificationPage> {
-    const query = cursor ? `?cursor=${encodeURIComponent(cursor)}` : "";
-    const body = await this.json<unknown>(`/notifications${query}`, {
-      headers: this.headers(),
-    });
-    const page = parseAgentNotificationPage(body);
-    if (!page) {
-      throw new Error("notification list response is invalid");
-    }
-    return page;
-  }
-
-  async notificationUnreadCount(): Promise<number> {
-    const body = await this.json<unknown>("/notifications/unread-count", {
-      headers: this.headers(),
-    });
-    if (
-      !body ||
-      typeof body !== "object" ||
-      !("unread" in body) ||
-      typeof (body as { unread: unknown }).unread !== "number"
-    ) {
-      throw new Error("notification unread-count response is invalid");
-    }
-    return (body as { unread: number }).unread;
-  }
-
-  async markNotificationsRead(ids: string[]): Promise<number> {
-    const body = await this.json<unknown>("/notifications/read", {
-      method: "POST",
-      headers: this.headers(),
-      body: JSON.stringify({ ids }),
-    });
-    return parseMarked(body);
-  }
-
-  async markAllNotificationsRead(): Promise<number> {
-    const body = await this.json<unknown>("/notifications/read-all", {
-      method: "POST",
-      headers: this.headers(),
-    });
-    return parseMarked(body);
-  }
-
-  async listInbox(): Promise<InboxEntry[]> {
-    const body = await this.json<unknown>("/inbox", {
-      headers: this.headers(),
-    });
-    if (!Array.isArray(body)) {
-      throw new Error("inbox response is not an array");
-    }
-    const entries: InboxEntry[] = [];
-    const seen = new Set<string>();
-    for (const value of body) {
-      const entry = parseInboxEntry(value);
-      if (!entry) {
-        throw new Error("inbox response contains invalid data");
-      }
-      const key = inboxConversationKey(entry.conversation);
-      if (seen.has(key)) {
-        throw new Error("inbox response lists a conversation twice");
-      }
-      seen.add(key);
-      entries.push(entry);
-    }
-    return entries;
-  }
-
-  async listPendingChatPrompts(): Promise<PendingChatPrompt[]> {
-    const body = await this.json<unknown>("/chats/pending-prompts", {
-      headers: this.headers(),
-    });
-    if (!Array.isArray(body)) {
-      throw new Error("pending chat prompt response is not an array");
-    }
-    const prompts = new Map<string, PendingChatPrompt>();
-    for (const item of body) {
-      const prompt = parsePendingChatPrompt(item);
-      if (!prompt || prompts.has(prompt.chatId)) {
-        throw new Error("pending chat prompt response contains invalid data");
-      }
-      prompts.set(prompt.chatId, prompt);
-    }
-    return [...prompts.values()];
-  }
-
-  deleteChat(chatId: string): Promise<void> {
-    return this.json(`/chats/${chatId}`, {
-      method: "DELETE",
-      headers: this.headers(),
-    });
-  }
-
-  listChatMessages(chatId: string): Promise<ChatTranscript> {
-    return this.json(`/chats/${chatId}/messages`, {
-      headers: this.headers(),
-    });
-  }
-
-  /**
-   * Compact this chat now, optionally saying what the summary should keep.
-   *
-   * Runs between turns only; the server refuses while one is running. A
-   * response with `compacted: false` is an ordinary answer — there was nothing
-   * worth summarizing — not a failure.
-   */
-  compactChat(chatId: string, focus?: string): Promise<CompactionRun> {
-    return this.json(`/chats/${encodeURIComponent(chatId)}/compact`, {
-      method: "POST",
-      headers: this.headers(true),
-      body: JSON.stringify(focus ? { focus } : {}),
-    });
-  }
-
-  undoTurnFileChanges(
-    chatId: string,
-    turnId: string,
-  ): Promise<{
-    chat_id: string;
-    turn_id: string;
-    files: ExecFileUndoOutcome[];
-  }> {
-    return this.json(
-      `/chats/${encodeURIComponent(chatId)}/turns/${encodeURIComponent(turnId)}/file-changes/undo`,
-      { method: "POST", headers: this.headers() },
-    );
-  }
-
-  undoFileChange(
-    chatId: string,
-    turnId: string,
-    snapshotId: string,
-  ): Promise<ExecFileUndoOutcome> {
-    return this.json(
-      `/chats/${encodeURIComponent(chatId)}/turns/${encodeURIComponent(turnId)}/file-changes/${encodeURIComponent(snapshotId)}/undo`,
-      { method: "POST", headers: this.headers() },
-    );
-  }
-
-  getFileChangePreview(
-    chatId: string,
-    turnId: string,
-    snapshotId: string,
-    revision: "before" | "after",
-    signal?: AbortSignal,
-  ): Promise<Blob> {
-    return this.blob(
-      `/chats/${encodeURIComponent(chatId)}/turns/${encodeURIComponent(turnId)}/file-changes/${encodeURIComponent(snapshotId)}/preview/${revision}`,
-      signal,
-    );
-  }
-
-  private async blob(path: string, signal?: AbortSignal): Promise<Blob> {
-    const response = await fetch(`${this.baseUrl}${path}`, {
-      headers: this.headers(),
-      signal,
-    });
-    await throwIfNotOk(response);
-    return response.blob();
-  }
-
-  /**
-   * Bytes read as they arrive, reporting how much has landed.
-   *
-   * Reading the body as a stream rather than awaiting it whole is the only way
-   * to say anything about a transfer while it is still running. It costs an
-   * extra copy — the chunks are joined once at the end — which is why the
-   * callers that have nothing to report progress to still use {@link blob}.
-   *
-   * `onProgress` is only ever called when the response declares its length:
-   * without a total there is no fraction to report, and a byte count climbing
-   * toward an unknown end is not worth a progress bar.
-   */
-  private async streamBytes(
-    path: string,
-    signal?: AbortSignal,
-    onProgress?: (progress: FileDownloadProgress) => void,
-  ): Promise<{ bytes: Uint8Array; contentType: string | null }> {
-    const response = await fetch(`${this.baseUrl}${path}`, {
-      headers: this.headers(),
-      signal,
-    });
-    await throwIfNotOk(response);
-
-    const contentType = response.headers.get("Content-Type");
-    const declared = Number(response.headers.get("Content-Length"));
-    const total = Number.isSafeInteger(declared) && declared > 0 ? declared : 0;
-
-    // No reader to stream from (an old runtime, or a mocked response in a
-    // test): take the whole body and skip straight to the finished state.
-    if (!response.body) {
-      const bytes = new Uint8Array(await response.arrayBuffer());
-      return { bytes, contentType };
-    }
-
-    const report =
-      onProgress && total > PROGRESS_MIN_BYTES ? throttle(onProgress) : null;
-    const reader = response.body.getReader();
-    const chunks: Uint8Array[] = [];
-    let loaded = 0;
-
-    while (true) {
-      const { done, value } = await reader.read();
-      if (done) break;
-      chunks.push(value);
-      loaded += value.length;
-      report?.({ loaded, total, percentage: (loaded / total) * 100 });
-    }
-
-    const bytes = new Uint8Array(loaded);
-    let offset = 0;
-    for (const chunk of chunks) {
-      bytes.set(chunk, offset);
-      offset += chunk.length;
-    }
-    return { bytes, contentType };
-  }
-
-  getChatImageAttachment(
-    chatId: string,
-    attachmentId: string,
-    signal?: AbortSignal,
-  ): Promise<Blob> {
-    return this.blob(
-      `/chats/${encodeURIComponent(chatId)}/attachments/images/${encodeURIComponent(attachmentId)}`,
-      signal,
-    );
-  }
-
-  /**
-   * Publish one file the renderer is already holding as a source on this chat.
-   *
-   * The native picker route reads the bytes in the host and imports them into
-   * the store inside this app. That is the wrong store while this window works
-   * on another machine, so a window in that state posts the bytes here instead
-   * and the machine that owns the conversation parses them.
-   *
-   * The media type decides which parser runs. A browser leaves `File.type`
-   * empty for a name it does not recognize, and the route refuses an empty
-   * `Content-Type`, so an unrecognized file is declared as opaque bytes and
-   * retained without extracted text rather than refused outright.
-   */
-  ingestChatDocument(
-    chatId: string,
-    file: File,
-  ): Promise<{ document_id: string }> {
-    return this.json(
-      `/chats/${encodeURIComponent(chatId)}/documents/raw?title=${encodeURIComponent(file.name)}`,
-      {
-        method: "POST",
-        headers: {
-          Authorization: `Bearer ${this.token}`,
-          "Content-Type": file.type || "application/octet-stream",
-        },
-        body: file,
-      },
-    );
-  }
-
-  /** One source's extracted text and catalog metadata. */
-  getChatDocument(chatId: string, documentId: string): Promise<DocumentDetail> {
-    return this.json(
-      `/chats/${encodeURIComponent(chatId)}/documents/${encodeURIComponent(documentId)}`,
-      { headers: this.headers() },
-    );
-  }
-
-  /**
-   * The original bytes of one source, exactly as they were imported.
-   *
-   * Bytes are addressed by document id inside its conversation and never by a
-   * host path, so a viewer can show the file the reader gave us without the
-   * renderer learning where on disk it came from. The conversation is part of
-   * the address rather than decoration: the server serves a document's bytes
-   * only under the chat that owns it.
-   *
-   * Returned as bytes rather than a URL because the renderer authenticates with
-   * a bearer header the webview cannot attach to an `<embed>` or `<img>` source,
-   * and because pdf.js and the workbook parsers want a buffer anyway. The
-   * stored media type comes back alongside them because it is what the text
-   * viewers dispatch on, and it would otherwise be lost when the streamed
-   * chunks are reassembled.
-   */
-  getChatDocumentFile(
-    chatId: string,
-    documentId: string,
-    signal?: AbortSignal,
-    onProgress?: (progress: FileDownloadProgress) => void,
-  ): Promise<{ bytes: Uint8Array; contentType: string | null }> {
-    return this.streamBytes(
-      `/chats/${encodeURIComponent(chatId)}/documents/${encodeURIComponent(documentId)}/file-content`,
-      signal,
-      onProgress,
-    );
-  }
-
-  patchChatTitle(chatId: string, title: string | null): Promise<Chat> {
-    return this.json(`/chats/${chatId}`, {
-      method: "PATCH",
-      headers: this.headers(true),
-      body: JSON.stringify({ title }),
-    });
-  }
-
-  patchChatModel(
-    chatId: string,
-    model: ModelSelectionKey | null,
-  ): Promise<Chat> {
-    return this.json(`/chats/${chatId}`, {
-      method: "PATCH",
-      headers: this.headers(true),
-      body: JSON.stringify({ model }),
-    });
-  }
-
-  patchChatReasoningEffort(
-    chatId: string,
-    reasoningEffort: ReasoningEffort | null,
-  ): Promise<Chat> {
-    return this.json(`/chats/${chatId}`, {
-      method: "PATCH",
-      headers: this.headers(true),
-      body: JSON.stringify({ reasoning_effort: reasoningEffort }),
-    });
-  }
-
-  patchChatPermissionMode(
-    chatId: string,
-    permissionMode: PermissionMode | null,
-  ): Promise<Chat> {
-    return this.json(`/chats/${chatId}`, {
-      method: "PATCH",
-      headers: this.headers(true),
-      body: JSON.stringify({ permission_mode: permissionMode }),
-    });
-  }
-
-  /**
-   * File a chat under a project, or take it back out with `null`.
-   *
-   * The server refuses (409) a chat that still holds connected folders: its
-   * folder grants are keyed to the identity it would be leaving.
-   */
-  moveChatToProject(chatId: string, projectId: string | null): Promise<Chat> {
-    return this.json(`/chats/${encodeURIComponent(chatId)}`, {
-      method: "PATCH",
-      headers: this.headers(true),
-      body: JSON.stringify({ project_id: projectId }),
-    });
-  }
-
-  patchChatNetworkPolicy(
-    chatId: string,
-    networkPolicy: NetworkPolicy,
-  ): Promise<Chat> {
-    return this.json(`/chats/${chatId}`, {
-      method: "PATCH",
-      headers: this.headers(true),
-      body: JSON.stringify({ network_policy: networkPolicy }),
-    });
-  }
-
-  /**
-   * The conversation's current task plan, or `null` when it has none.
-   *
-   * The journal only carries a hint that the plan moved on, so this is where
-   * the steps come from — on the hint, and again on reload. The payload is
-   * model-authored text, so it is validated here rather than trusted.
-   */
-  async getTaskPlan(chatId: string): Promise<TaskPlan | null> {
-    const body = await this.json<unknown>(
-      `/chats/${encodeURIComponent(chatId)}/task-plan`,
-      { headers: this.headers() },
-    );
-    return parseTaskPlan(body);
-  }
-
-  listAgentRuns(chatId: string): Promise<AgentRun[]> {
-    return this.json(`/chats/${chatId}/agent-runs`, {
-      headers: this.headers(),
-    });
-  }
-
-  /**
-   * The ordered, renderer-safe activity history for one background run.
-   *
-   * Malformed or unknown entries are dropped rather than trusted, keeping the
-   * closed vocabulary the server promises. A wrong-chat, foreground, or missing
-   * run answers `404`, which surfaces as a thrown error.
-   */
-  async listAgentRunActivity(
-    chatId: string,
-    runId: string,
-  ): Promise<AgentActivityHistoryEntry[]> {
-    const body = await this.json<unknown>(
-      `/chats/${encodeURIComponent(chatId)}/agent-runs/${encodeURIComponent(runId)}/activity`,
-      { headers: this.headers() },
-    );
-    return parseAgentActivityHistory(body);
-  }
-
-  /**
-   * The full ordered checklist one background run keeps, or `null`.
-   *
-   * The run snapshot already carries the count and the current step, which is
-   * all a status row needs; this is the list behind it, read when a reader
-   * opens the run. A wrong-chat, foreground, or missing run answers `404`,
-   * which surfaces as a thrown error.
-   */
-  async getAgentRunTaskPlan(
-    chatId: string,
-    runId: string,
-  ): Promise<AgentRunTaskPlan | null> {
-    const body = await this.json<unknown>(
-      `/chats/${encodeURIComponent(chatId)}/agent-runs/${encodeURIComponent(runId)}/task-plan`,
-      { headers: this.headers() },
-    );
-    return parseAgentRunTaskPlan(body);
-  }
-
-  /**
-   * One resumable page of a background run's live progress.
-   *
-   * Poll with the previous page's `nextSequence` to receive only what the run
-   * has published since. A wrong-chat, foreground, or missing run answers
-   * `404`, which surfaces as a thrown error.
-   */
-  async listAgentRunProgress(
-    chatId: string,
-    runId: string,
-    afterSequence = 0,
-    limit?: number,
-  ): Promise<AgentRunProgress> {
-    const query = new URLSearchParams({
-      after_sequence: String(afterSequence),
-    });
-    if (limit !== undefined) query.set("limit", String(limit));
-    const body = await this.json<unknown>(
-      `/chats/${encodeURIComponent(chatId)}/agent-runs/${encodeURIComponent(runId)}/progress?${query}`,
-      { headers: this.headers() },
-    );
-    return parseAgentRunProgress(body, afterSequence);
-  }
-
-  /** Resume a background run paused at a check-in, optionally with guidance. */
-  async resumeAgentRun(
-    chatId: string,
-    runId: string,
-    guidance?: string,
-  ): Promise<void> {
-    await this.json<unknown>(
-      `/chats/${encodeURIComponent(chatId)}/agent-runs/${encodeURIComponent(runId)}/resume`,
-      {
-        method: "POST",
-        headers: this.headers(true),
-        body: JSON.stringify(guidance ? { guidance } : {}),
-      },
-      202,
-    );
-  }
-
-  async cancelAgentRun(
-    chatId: string,
-    runId: string,
-  ): Promise<SandboxAgentCancellation> {
-    const body = await this.json<unknown>(
-      `/chats/${encodeURIComponent(chatId)}/agent-runs/${encodeURIComponent(runId)}/cancel`,
-      {
-        method: "POST",
-        headers: this.headers(),
-      },
-      202,
-    );
-    const cancellation = parseSandboxAgentCancellation(body);
-    if (!cancellation || cancellation.id !== runId) {
-      throw new Error("sandbox cancellation response is invalid");
-    }
-    return cancellation;
-  }
-
-  /**
-   * `attachments` names images already published for this chat, in the order
-   * they should be shown to the model. Only identity crosses: the server
-   * re-derives every attachment's format and dimensions from the stored bytes.
-   *
-   * `invokedSkills` names the skills the reader explicitly reached for. A name
-   * the install cannot run refuses the whole turn rather than being dropped, so
-   * the caller must be ready to show the refusal.
-   */
-  postMessage(
-    chatId: string,
-    turnId: string,
-    content: string,
-    attachments: readonly string[] = [],
-    fileAttachments: readonly string[] = [],
-    invokedSkills: readonly string[] = [],
-    voiceInputUsed = false,
-    queue = false,
-  ): Promise<void> {
-    return this.json(`/chats/${chatId}/messages`, {
-      method: "POST",
-      headers: this.headers(true),
-      body: JSON.stringify({
-        turn_id: turnId,
-        content,
-        attachments,
-        file_attachments: fileAttachments,
-        invoked_skills: invokedSkills,
-        voice_input_used: voiceInputUsed,
-        queue,
-      }),
-    });
-  }
-
-  /** The chat's queued messages, FIFO, plus whether promotion is paused. */
-  listQueuedTurns(
-    chatId: string,
-  ): Promise<{ queued: QueuedTurn[]; paused: boolean }> {
-    return this.json(`/chats/${encodeURIComponent(chatId)}/queued`, {
-      headers: this.headers(),
-    });
-  }
-
-  patchQueuedTurn(
-    chatId: string,
-    turnId: string,
-    update: { content?: string; position?: number },
-  ): Promise<QueuedTurn> {
-    return this.json(
-      `/chats/${encodeURIComponent(chatId)}/queued/${encodeURIComponent(turnId)}`,
-      {
-        method: "PATCH",
-        headers: this.headers(true),
-        body: JSON.stringify(update),
-      },
-    );
-  }
-
-  async deleteQueuedTurn(chatId: string, turnId: string): Promise<void> {
-    await this.json<unknown>(
-      `/chats/${encodeURIComponent(chatId)}/queued/${encodeURIComponent(turnId)}`,
-      { method: "DELETE", headers: this.headers() },
-      204,
-    );
-  }
-
-  async putQueuePaused(chatId: string, paused: boolean): Promise<void> {
-    await this.json<unknown>(
-      `/chats/${encodeURIComponent(chatId)}/queue-paused`,
-      {
-        method: "PUT",
-        headers: this.headers(true),
-        body: JSON.stringify({ paused }),
-      },
-      204,
-    );
-  }
-
-  /** Release a paused queue so the oldest message starts on the next sweep. */
-  async sendQueuedNow(chatId: string): Promise<void> {
-    await this.json<unknown>(
-      `/chats/${encodeURIComponent(chatId)}/queued/send-now`,
-      { method: "POST", headers: this.headers() },
-      204,
-    );
-  }
-
-  steer(
-    chatId: string,
-    turnId: string,
-    steerId: string,
-    content: string,
-    interrupt = false,
-    voiceInputUsed = false,
-    invokedSkills: readonly string[] = [],
-  ): Promise<void> {
-    return this.json(`/chats/${chatId}/steer`, {
-      method: "POST",
-      headers: this.headers(true),
-      body: JSON.stringify({
-        steer_id: steerId,
-        turn_id: turnId,
-        content,
-        interrupt,
-        voice_input_used: voiceInputUsed,
-        invoked_skills: invokedSkills,
-      }),
-    });
-  }
-
-  cancel(chatId: string, turnId: string): Promise<void> {
-    return this.json(`/chats/${chatId}/cancel`, {
-      method: "POST",
-      headers: this.headers(true),
-      body: JSON.stringify({ turn_id: turnId }),
-    });
-  }
-
-  decideApproval(
-    chatId: string,
-    callId: string,
-    decision: "approve" | "reject",
-    grant: ApprovalGrantRung | null = null,
-  ): Promise<void> {
-    return this.json(`/chats/${chatId}/approvals/${callId}`, {
-      method: "POST",
-      headers: this.headers(true),
-      body: JSON.stringify({ decision, grant }),
-    });
-  }
-
-  async listPendingApprovals(chatId: string): Promise<PendingToolApproval[]> {
-    const body = await this.json<unknown>(`/chats/${chatId}/approvals`, {
-      headers: this.headers(),
-    });
-    if (!Array.isArray(body)) {
-      throw new Error("pending approval response is not an array");
-    }
-
-    const approvals = new Map<string, PendingToolApproval>();
-    let turnId: string | null = null;
-    for (const item of body) {
-      const approval = parsePendingToolApproval(item);
-      if (!approval) {
-        throw new Error("pending approval response contains an invalid item");
-      }
-      if (approvals.has(approval.callId)) {
-        throw new Error("pending approval response contains a duplicate call");
-      }
-      if (turnId !== null && turnId !== approval.turnId) {
-        throw new Error("pending approval response spans multiple turns");
-      }
-      turnId = approval.turnId;
-      approvals.set(approval.callId, approval);
-    }
-    return [...approvals.values()];
-  }
-
-  /** Every standing "don't ask again", newest first, across all chats. */
-  listStandingGrants(): Promise<StandingGrantSnapshot[]> {
-    return this.json(`/grants`, { headers: this.headers() });
-  }
-
-  /**
-   * The server's rows of the unified consent read model: every standing tool
-   * grant as one consent statement. The capability half comes from the host
-   * broker over the Tauri boundary and joins these rows renderer-side.
-   */
-  listConsentStatements(): Promise<ConsentStatementSnapshot[]> {
-    return this.json(`/consent/statements`, { headers: this.headers() });
-  }
-
-  /** Withdraw a standing grant; later matching calls ask again. */
-  revokeStandingGrant(sourceCallId: string): Promise<void> {
-    return this.json(`/grants/${sourceCallId}`, {
-      method: "DELETE",
-      headers: this.headers(),
-    });
-  }
-
-  async listPendingFolderAccessRequests(
-    chatId: string,
-  ): Promise<PendingFolderAccessRequest[]> {
-    const body = await this.json<unknown>(
-      `/chats/${chatId}/client-executions/pending`,
-      { headers: this.headers() },
-    );
-    if (!Array.isArray(body)) return [];
-
-    const requests = new Map<string, PendingFolderAccessRequest>();
-    for (const item of body) {
-      const request = parseFolderAccessRequest(item);
-      if (request && !requests.has(request.callId)) {
-        requests.set(request.callId, request);
-      }
-    }
-    return [...requests.values()];
-  }
-
-  async listPendingOutputWritebackRequests(
-    chatId: string,
-  ): Promise<PendingOutputWritebackRequest[]> {
-    const body = await this.json<unknown>(
-      `/chats/${chatId}/output-writebacks/pending`,
-      { headers: this.headers() },
-    );
-    if (!Array.isArray(body)) {
-      throw new Error("pending output write-back response is not an array");
-    }
-
-    const requests = new Map<string, PendingOutputWritebackRequest>();
-    for (const item of body) {
-      const request = parseOutputWritebackRequest(item);
-      if (!request || requests.has(request.callId)) {
-        throw new Error(
-          "pending output write-back response contains invalid data",
-        );
-      }
-      requests.set(request.callId, request);
-    }
-    return [...requests.values()];
-  }
-
-  async listPendingUserQuestions(
-    chatId: string,
-  ): Promise<PendingUserQuestions[]> {
-    const body = await this.json<unknown>(
-      `/chats/${chatId}/questions/pending`,
-      {
-        headers: this.headers(),
-      },
-    );
-    if (!Array.isArray(body)) {
-      throw new Error("pending question response is not an array");
-    }
-    const requests = new Map<string, PendingUserQuestions>();
-    for (const item of body) {
-      const request = parsePendingUserQuestions(item);
-      if (!request || requests.has(request.callId)) {
-        throw new Error("pending question response contains invalid data");
-      }
-      requests.set(request.callId, request);
-    }
-    return [...requests.values()];
-  }
-
-  async answerUserQuestions(
-    chatId: string,
-    callId: string,
-    answers: UserQuestionAnswer[],
-    additionalUserContext?: string,
-  ): Promise<void> {
-    await this.json(`/chats/${chatId}/questions/${callId}/answer`, {
-      method: "POST",
-      headers: this.headers(true),
-      body: JSON.stringify({
-        answers: answers.map((answer) => ({
-          question_id: answer.questionId,
-          selected_option_ids: answer.selectedOptionIds,
-          ...(answer.customAnswer === undefined
-            ? {}
-            : { custom_answer: answer.customAnswer }),
-        })),
-        ...(additionalUserContext === undefined
-          ? {}
-          : { additional_user_context: additionalUserContext }),
-      }),
-    });
-  }
-
-  async listPendingPlanApprovals(
-    chatId: string,
-  ): Promise<PendingPlanApproval[]> {
-    const body = await this.json<unknown>(`/chats/${chatId}/plans/pending`, {
-      headers: this.headers(),
-    });
-    if (!Array.isArray(body)) {
-      throw new Error("pending plan response is not an array");
-    }
-    const requests = new Map<string, PendingPlanApproval>();
-    for (const item of body) {
-      const request = parsePendingPlanApproval(item);
-      if (!request || requests.has(request.callId)) {
-        throw new Error("pending plan response contains invalid data");
-      }
-      requests.set(request.callId, request);
-    }
-    return [...requests.values()];
-  }
-
-  async decidePlan(
-    chatId: string,
-    callId: string,
-    decision: PlanDecision,
-  ): Promise<void> {
-    await this.json(`/chats/${chatId}/plans/${callId}/decision`, {
-      method: "POST",
-      headers: this.headers(true),
-      body: JSON.stringify(
-        decision.decision === "reject" && decision.feedback !== undefined
-          ? { decision: "reject", feedback: decision.feedback }
-          : { decision: decision.decision },
-      ),
-    });
-  }
-
-  /** Open the chat event stream; auth via Sec-WebSocket-Protocol. */
-  openEvents(
-    chatId: string,
-    after: number,
-    onFrame: (frame: ChatFrame) => void,
-  ): WebSocket {
-    const url = `${this.baseUrl.replace(/^http/, "ws")}/chats/${chatId}/events?after=${after}`;
-    const protocols = [WS_HANDSHAKE, `${WS_TOKEN_PREFIX}${this.token}`];
-    const socket = new WebSocket(url, protocols);
-    socket.onmessage = (msg) => {
-      try {
-        onFrame(JSON.parse(String(msg.data)) as ChatFrame);
-      } catch (err) {
-        console.error("bad event frame", err);
-      }
-    };
-    return socket;
-  }
-
-  async getCodeDeliveryRepositories(
-    options?: DeliveryRequestOptions,
-  ): Promise<CodeDeliveryRepositoriesSnapshot> {
-    const path = options?.refreshAuth
-      ? "/code/delivery/repositories?refresh=true"
-      : "/code/delivery/repositories";
-    return requireParsed(
-      parseCodeDeliveryRepositories(
-        await this.deliveryJson(
-          path,
-          {
-            headers: this.headers(),
-          },
-          options,
+export class ApiClient extends withCodeEventsApi(
+  withCodeTerminalsApi(
+    withCodeGrantsApi(
+      withCodeGitApi(
+        withCodeFilesApi(
+          withCodeSessionsApi(
+            withCodeWorkspacesApi(
+              withCodeReposApi(
+                withDeliveryApi(
+                  withTurnsApi(
+                    withAgentRunsApi(
+                      withChatApi(
+                        withProjectsApi(withAppsApi(withSettingsApi(HttpCore))),
+                      ),
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          ),
         ),
       ),
-      "code delivery repositories",
-    );
+    ),
+  ),
+) {
+  constructor(baseUrl: string, token: string) {
+    super(baseUrl, token);
   }
-
-  async resolveCodeDeliveryRepositories(
-    repositories: string[],
-    options?: DeliveryRequestOptions,
-  ): Promise<CodeDeliveryRepositoriesSnapshot> {
-    return requireParsed(
-      parseCodeDeliveryRepositories(
-        await this.deliveryJson(
-          "/code/delivery/repositories/resolve",
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify({ repositories }),
-          },
-          options,
-        ),
-      ),
-      "code delivery repositories",
-    );
-  }
-
-  async queryCodeDeliveryPullRequests(
-    query: CodeDeliveryPullRequestQuery,
-    options?: DeliveryRequestOptions,
-  ): Promise<CodeDeliveryPullRequestsPage> {
-    return requireParsed(
-      parseCodeDeliveryPullRequestsPage(
-        await this.deliveryJson(
-          "/code/delivery/pull-requests/query",
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify(query),
-          },
-          options,
-        ),
-      ),
-      "code delivery pull requests",
-    );
-  }
-
-  async getCodeDeliveryPullRequestDetail(
-    target: CodeDeliveryPullRequestTarget,
-    options?: DeliveryRequestOptions,
-  ): Promise<CodeDeliveryPullRequestDetail> {
-    return requireParsed(
-      parseCodeDeliveryPullRequestDetail(
-        await this.deliveryJson(
-          "/code/delivery/pull-requests/detail",
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify(target),
-          },
-          options,
-        ),
-      ),
-      "code delivery pull request detail",
-    );
-  }
-
-  async runCodeDeliveryPullRequestAction(
-    body: CodeDeliveryPullRequestActionBody,
-    options?: DeliveryRequestOptions,
-  ): Promise<CodeDeliveryActionResult> {
-    return requireParsed(
-      parseCodeDeliveryActionResult(
-        await this.deliveryJson(
-          "/code/delivery/pull-requests/action",
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify(body),
-          },
-          options,
-        ),
-      ),
-      "code delivery pull request action",
-    );
-  }
-
-  async queryCodeDeliveryRuns(
-    query: CodeDeliveryRunQuery,
-    options?: DeliveryRequestOptions,
-  ): Promise<CodeDeliveryRunsPage> {
-    return requireParsed(
-      parseCodeDeliveryRunsPage(
-        await this.deliveryJson(
-          "/code/delivery/runs/query",
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify(query),
-          },
-          options,
-        ),
-      ),
-      "code delivery runs",
-    );
-  }
-
-  async getCodeDeliveryRunDetail(
-    target: CodeDeliveryRunTarget,
-    options?: DeliveryRequestOptions,
-  ): Promise<CodeDeliveryRunDetail> {
-    return requireParsed(
-      parseCodeDeliveryRunDetail(
-        await this.deliveryJson(
-          "/code/delivery/runs/detail",
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify(target),
-          },
-          options,
-        ),
-      ),
-      "code delivery run detail",
-    );
-  }
-
-  async runCodeDeliveryRunAction(
-    body: CodeDeliveryRunActionBody,
-    options?: DeliveryRequestOptions,
-  ): Promise<CodeDeliveryActionResult> {
-    return requireParsed(
-      parseCodeDeliveryActionResult(
-        await this.deliveryJson(
-          "/code/delivery/runs/action",
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify(body),
-          },
-          options,
-        ),
-      ),
-      "code delivery run action",
-    );
-  }
-
-  async listCodeRepos(): Promise<CodeRepoSnapshot[]> {
-    const body = await this.json<unknown>("/code/repos", {
-      headers: this.headers(),
-    });
-    return parseList(body, parseCodeRepo, "code repos");
-  }
-
-  async createCodeRepo(body: {
-    path: string;
-    display_name?: string;
-    default_base_ref?: string;
-    branch_prefix?: string;
-    setup_script?: string;
-    archive_script?: string;
-    quick_actions?: QuickAction[];
-  }): Promise<CodeRepoSnapshot> {
-    return requireParsed(
-      parseCodeRepo(
-        await this.json("/code/repos", {
-          method: "POST",
-          headers: this.headers(true),
-          body: JSON.stringify(body),
-        }),
-      ),
-      "code repo",
-    );
-  }
-
-  async getCodeRepo(repoId: string): Promise<CodeRepoSnapshot> {
-    return requireParsed(
-      parseCodeRepo(
-        await this.json(`/code/repos/${encodeURIComponent(repoId)}`, {
-          headers: this.headers(),
-        }),
-      ),
-      "code repo",
-    );
-  }
-
-  async patchCodeRepo(
-    repoId: string,
-    body: {
-      display_name?: string;
-      default_base_ref?: string;
-      branch_prefix?: string;
-      setup_script?: string | null;
-      archive_script?: string | null;
-      /** The whole list, replaced. Omit to leave the stored list alone. */
-      quick_actions?: QuickAction[];
-    },
-  ): Promise<CodeRepoSnapshot> {
-    return requireParsed(
-      parseCodeRepo(
-        await this.json(`/code/repos/${encodeURIComponent(repoId)}`, {
-          method: "PATCH",
-          headers: this.headers(true),
-          body: JSON.stringify(body),
-        }),
-      ),
-      "code repo",
-    );
-  }
-
-  deleteCodeRepo(repoId: string): Promise<void> {
-    return this.json(`/code/repos/${encodeURIComponent(repoId)}`, {
-      method: "DELETE",
-      headers: this.headers(),
-    });
-  }
-
-  /**
-   * What this machine can add a repository from.
-   *
-   * Member-plane, unlike `getCodeCloneDefaults`, which is administrator-only:
-   * on a shared machine the person adding a repository is usually not an
-   * administrator, and a dialog that cannot read its own options is no dialog.
-   */
-  async listCodeGithubRepositories(
-    signal?: AbortSignal,
-  ): Promise<CodeGithubRepositories> {
-    return requireParsed(
-      parseCodeGithubRepositories(
-        await this.json("/code/repos/github", {
-          headers: this.headers(),
-          signal,
-        }),
-      ),
-      "github repositories",
-    );
-  }
-
-  async getCodeRepoSources(signal?: AbortSignal): Promise<CodeRepoSources> {
-    return requireParsed(
-      parseCodeRepoSources(
-        await this.json("/code/repos/sources", {
-          headers: this.headers(),
-          signal,
-        }),
-      ),
-      "repo sources",
-    );
-  }
-
-  async getCodeCloneDefaults(signal?: AbortSignal): Promise<CodeCloneDefaults> {
-    return requireParsed(
-      parseCodeCloneDefaults(
-        await this.json("/code/repos/clone-defaults", {
-          headers: this.headers(),
-          signal,
-        }),
-      ),
-      "clone defaults",
-    );
-  }
-
-  /**
-   * Start a clone. `parent_dir` is omitted when the machine places clones
-   * itself — see `CodeRepoSources.chooses_destination`.
-   */
-  async startCodeClone(body: {
-    url?: string;
-    github?: string;
-    parent_dir?: string;
-    name?: string;
-  }): Promise<CodeCloneJobSnapshot> {
-    return requireParsed(
-      parseCodeCloneJob(
-        await this.json("/code/repos/clone", {
-          method: "POST",
-          headers: this.headers(true),
-          body: JSON.stringify(body),
-        }),
-      ),
-      "clone job",
-    );
-  }
-
-  async getCodeWorktreeRoot(): Promise<CodeWorktreeRoot> {
-    return requireParsed(
-      parseCodeWorktreeRoot(
-        await this.json("/code/worktree-root", { headers: this.headers() }),
-      ),
-      "worktree root",
-    );
-  }
-
-  /**
-   * Move the root new worktrees are created under. A null root restores the
-   * default. Worktrees already on disk keep the path they were created at.
-   */
-  async setCodeWorktreeRoot(root: string | null): Promise<CodeWorktreeRoot> {
-    return requireParsed(
-      parseCodeWorktreeRoot(
-        await this.json("/code/worktree-root", {
-          method: "PUT",
-          headers: this.headers(true),
-          body: JSON.stringify({ root }),
-        }),
-      ),
-      "worktree root",
-    );
-  }
-
-  async getCodeCloneJob(jobId: string): Promise<CodeCloneJobSnapshot> {
-    return requireParsed(
-      parseCodeCloneJob(
-        await this.json(`/code/repos/clone/${encodeURIComponent(jobId)}`, {
-          headers: this.headers(),
-        }),
-      ),
-      "clone job",
-    );
-  }
-
-  async getHarnessDoctor(): Promise<HarnessDoctorReport> {
-    return requireParsed(
-      parseHarnessDoctorReport(
-        await this.json("/code/harnesses", { headers: this.headers() }),
-      ),
-      "harness doctor",
-    );
-  }
-
-  async getCodeSubscriptionUsage(): Promise<CodeSubscriptionUsage> {
-    return requireParsed(
-      parseCodeSubscriptionUsage(
-        await this.json("/code/usage", { headers: this.headers() }),
-      ),
-      "subscription usage",
-    );
-  }
-
-  async getCodeAnalytics(
-    range: CodeAnalyticsRange,
-    repoId?: string,
-  ): Promise<CodeAnalyticsSnapshot> {
-    const query = new URLSearchParams({ range });
-    if (repoId) query.set("repo_id", repoId);
-    return requireParsed(
-      parseCodeAnalytics(
-        await this.json(`/code/analytics?${query.toString()}`, {
-          headers: this.headers(),
-        }),
-      ),
-      "code analytics",
-    );
-  }
-
-  async listCodeHarnessModels(
-    kind: HarnessKind,
-  ): Promise<ParsedHarnessModelList> {
-    return requireParsed(
-      parseHarnessModelList(
-        await this.json(`/code/harnesses/${encodeURIComponent(kind)}/models`, {
-          headers: this.headers(),
-        }),
-      ),
-      "harness models",
-    );
-  }
-
-  /**
-   * Warm the pinned install of one engine ahead of a session create.
-   *
-   * Answers as soon as the server knows where the install stands; the phases
-   * that follow arrive on `WS /code/updates`. Safe to call repeatedly — an
-   * installed pin answers `ready` and one already running is not restarted.
-   */
-  /**
-   * Start — or report — the pinned download of one engine.
-   *
-   * `deliberate` marks a reader who pressed Install rather than a picker
-   * warming its selection. Only that case retries a managed-Node install that
-   * already failed.
-   */
-  async startHarnessInstall(
-    kind: HarnessKind,
-    deliberate = false,
-  ): Promise<CodeHarnessInstallSnapshot> {
-    const path = `/code/harnesses/${encodeURIComponent(kind)}/install`;
-    return requireParsed(
-      parseCodeHarnessInstall(
-        await this.json(deliberate ? `${path}?deliberate=true` : path, {
-          method: "POST",
-          headers: this.headers(),
-        }),
-      ),
-      "harness install",
-    );
-  }
-
-  async refreshHarnessDoctor(): Promise<HarnessDoctorReport> {
-    return requireParsed(
-      parseHarnessDoctorReport(
-        await this.json("/code/harnesses/refresh", {
-          method: "POST",
-          headers: this.headers(),
-        }),
-      ),
-      "harness doctor",
-    );
-  }
-
-  async listCodeWorkspaces(repoId?: string): Promise<CodeWorkspaceSnapshot[]> {
-    const query = repoId ? `?repo_id=${encodeURIComponent(repoId)}` : "";
-    const body = await this.json<unknown>(`/code/workspaces${query}`, {
-      headers: this.headers(),
-    });
-    return parseList(body, parseCodeWorkspace, "code workspaces");
-  }
-
-  async createCodeWorkspace(body: {
-    repo_id: string;
-    title?: string;
-    base_ref?: string;
-  }): Promise<CodeWorkspaceSnapshot> {
-    return requireParsed(
-      parseCodeWorkspace(
-        await this.json("/code/workspaces", {
-          method: "POST",
-          headers: this.headers(true),
-          body: JSON.stringify(body),
-        }),
-      ),
-      "code workspace",
-    );
-  }
-
-  async getCodeWorkspace(workspaceId: string): Promise<CodeWorkspaceSnapshot> {
-    return requireParsed(
-      parseCodeWorkspace(
-        await this.json(`/code/workspaces/${encodeURIComponent(workspaceId)}`, {
-          headers: this.headers(),
-        }),
-      ),
-      "code workspace",
-    );
-  }
-
-  async patchCodeWorkspace(
-    workspaceId: string,
-    body: { title: string },
-  ): Promise<CodeWorkspaceSnapshot> {
-    return requireParsed(
-      parseCodeWorkspace(
-        await this.json(`/code/workspaces/${encodeURIComponent(workspaceId)}`, {
-          method: "PATCH",
-          headers: this.headers(true),
-          body: JSON.stringify(body),
-        }),
-      ),
-      "code workspace",
-    );
-  }
-
-  async archiveCodeWorkspace(
-    workspaceId: string,
-    force = false,
-  ): Promise<CodeWorkspaceSnapshot> {
-    return requireParsed(
-      parseCodeWorkspace(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/archive`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify({ force }),
-          },
-        ),
-      ),
-      "code workspace",
-    );
-  }
-
-  /**
-   * Reactivate an archived workspace: worktree back at its path, on its kept
-   * branch. 409 kinds `branch_missing` and `worktree_path_occupied` mean a
-   * true restore is impossible; the caller offers the new-workspace fallback.
-   */
-  async restoreCodeWorkspace(
-    workspaceId: string,
-  ): Promise<CodeWorkspaceSnapshot> {
-    return requireParsed(
-      parseCodeWorkspace(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/restore`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-          },
-        ),
-      ),
-      "code workspace",
-    );
-  }
-
-  /**
-   * Run the repo's setup script again on a workspace that is `setup_failed`.
-   * The worktree is the one it already has, so a success takes the workspace
-   * Active without cutting a second checkout. A second failure comes back as
-   * 422 `setup_failed`.
-   */
-  async retryCodeWorkspaceSetup(
-    workspaceId: string,
-  ): Promise<CodeWorkspaceSnapshot> {
-    return requireParsed(
-      parseCodeWorkspace(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/retry-setup`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-          },
-        ),
-      ),
-      "code workspace",
-    );
-  }
-
-  async listCodeWorkspaceSessions(
-    workspaceId: string,
-  ): Promise<CodeSessionSnapshot[]> {
-    const body = await this.json<unknown>(
-      `/code/workspaces/${encodeURIComponent(workspaceId)}/sessions`,
-      { headers: this.headers() },
-    );
-    return requireParsed(parseCodeSessionList(body), "code workspace sessions");
-  }
-
-  async createCodeSession(
-    workspaceId: string,
-    body: {
-      harness: HarnessKind;
-      permission_mode: PermissionMode;
-      model?: string;
-      reasoning_effort?: ReasoningEffort;
-      fast_mode?: boolean;
-    },
-  ): Promise<CodeSessionSnapshot> {
-    return requireParsed(
-      parseCodeSession(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/sessions`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify(body),
-          },
-        ),
-      ),
-      "code session",
-    );
-  }
-
-  async getCodeSessionDebug(sessionId: string): Promise<unknown> {
-    return this.json(`/code/sessions/${encodeURIComponent(sessionId)}/debug`, {
-      headers: this.headers(),
-    });
-  }
-
-  async listCodeSessionTurns(sessionId: string): Promise<CodeTurnSnapshot[]> {
-    const body = await this.json<unknown>(
-      `/code/sessions/${encodeURIComponent(sessionId)}/turns`,
-      { headers: this.headers() },
-    );
-    return requireParsed(parseCodeTurnList(body), "code session turns");
-  }
-
-  /**
-   * Submit a turn, or park it behind the one in flight.
-   *
-   * The route answers 202 either way: a turn snapshot when the session was
-   * idle, a queue receipt when it was busy. Both are accepted work — treating
-   * the receipt as a malformed turn would report a failure for a message the
-   * server holds, and a retry would double-send.
-   */
-  async submitCodeTurn(
-    sessionId: string,
-    message: string,
-    model?: string,
-    attachments?: readonly { blob_id: string; media_type: string }[],
-    /**
-     * Omit to leave the session's stored level alone. Pass `null` to hand the
-     * level back to the engine's own default — that is a choice, not an
-     * omission, so it cannot ride on `undefined`.
-     */
-    reasoningEffort?: ReasoningEffort | null,
-  ): Promise<CodeTurnSubmission> {
-    return requireParsed(
-      parseCodeTurnSubmission(
-        await this.json(
-          `/code/sessions/${encodeURIComponent(sessionId)}/turns`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify({
-              message,
-              ...(model ? { model } : {}),
-              ...(reasoningEffort !== undefined
-                ? { reasoning_effort: reasoningEffort }
-                : {}),
-              ...(attachments && attachments.length > 0 ? { attachments } : {}),
-            }),
-          },
-        ),
-      ),
-      "code turn",
-    );
-  }
-
-  /**
-   * Move a live session onto a different permission mode.
-   *
-   * The engine re-postures in place where it can and is relaunched where it
-   * cannot; either way the new mode governs from the next turn. Refused (409
-   * `turn_running`) while a turn is in flight.
-   */
-  async setCodeSessionPermissionMode(
-    sessionId: string,
-    permissionMode: PermissionMode,
-  ): Promise<CodeSessionSnapshot> {
-    return requireParsed(
-      parseCodeSession(
-        await this.json(
-          `/code/sessions/${encodeURIComponent(sessionId)}/mode`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify({ permission_mode: permissionMode }),
-          },
-        ),
-      ),
-      "code session",
-    );
-  }
-
-  /** Change a session's reasoning effort. `null` is the engine's own default. */
-  async setCodeSessionReasoningEffort(
-    sessionId: string,
-    reasoningEffort: ReasoningEffort | null,
-  ): Promise<CodeSessionSnapshot> {
-    return requireParsed(
-      parseCodeSession(
-        await this.json(
-          `/code/sessions/${encodeURIComponent(sessionId)}/effort`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify({ reasoning_effort: reasoningEffort }),
-          },
-        ),
-      ),
-      "code session",
-    );
-  }
-
-  /** Arm or disarm the engine's fast mode for a session. */
-  async setCodeSessionFastMode(
-    sessionId: string,
-    fastMode: boolean,
-  ): Promise<CodeSessionSnapshot> {
-    return requireParsed(
-      parseCodeSession(
-        await this.json(
-          `/code/sessions/${encodeURIComponent(sessionId)}/fast-mode`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify({ fast_mode: fastMode }),
-          },
-        ),
-      ),
-      "code session",
-    );
-  }
-
-  getCodeSessionImage(
-    sessionId: string,
-    blobId: string,
-    signal?: AbortSignal,
-  ): Promise<Blob> {
-    return this.blob(
-      `/code/sessions/${encodeURIComponent(sessionId)}/attachments/images/${encodeURIComponent(blobId)}`,
-      signal,
-    );
-  }
-
-  /**
-   * Redirect the in-flight turn. The route answers with an empty status;
-   * unsupported engines refuse with `steering_unavailable` rather than queue.
-   */
-  steerCodeSession(
-    sessionId: string,
-    expectedTurnId: string,
-    guidance: string,
-  ): Promise<void> {
-    return this.json(`/code/sessions/${encodeURIComponent(sessionId)}/steer`, {
-      method: "POST",
-      headers: this.headers(true),
-      body: JSON.stringify({
-        expected_turn_id: expectedTurnId,
-        guidance,
-      }),
-    });
-  }
-
-  interruptCodeSession(sessionId: string): Promise<void> {
-    return this.json(
-      `/code/sessions/${encodeURIComponent(sessionId)}/interrupt`,
-      {
-        method: "POST",
-        headers: this.headers(),
-      },
-    );
-  }
-
-  /** The session's queued messages, FIFO, plus whether promotion is paused. */
-  async listCodeQueuedTurns(
-    sessionId: string,
-  ): Promise<{ queued: QueuedCodeTurn[]; paused: boolean }> {
-    const snapshot = await this.json<{ queued: unknown[]; paused: boolean }>(
-      `/code/sessions/${encodeURIComponent(sessionId)}/queued`,
-      { headers: this.headers() },
-    );
-    return {
-      queued: snapshot.queued.map((row) =>
-        requireParsed(parseQueuedCodeTurn(row), "queued code turn"),
-      ),
-      paused: snapshot.paused === true,
-    };
-  }
-
-  async patchCodeQueuedTurn(
-    sessionId: string,
-    queuedId: string,
-    update: { message?: string; position?: number },
-  ): Promise<QueuedCodeTurn> {
-    return requireParsed(
-      parseQueuedCodeTurn(
-        await this.json(
-          `/code/sessions/${encodeURIComponent(sessionId)}/queued/${encodeURIComponent(queuedId)}`,
-          {
-            method: "PATCH",
-            headers: this.headers(true),
-            body: JSON.stringify(update),
-          },
-        ),
-      ),
-      "queued code turn",
-    );
-  }
-
-  async deleteCodeQueuedTurn(
-    sessionId: string,
-    queuedId: string,
-  ): Promise<void> {
-    await this.json<unknown>(
-      `/code/sessions/${encodeURIComponent(sessionId)}/queued/${encodeURIComponent(queuedId)}`,
-      { method: "DELETE", headers: this.headers() },
-      204,
-    );
-  }
-
-  async putCodeQueuePaused(sessionId: string, paused: boolean): Promise<void> {
-    await this.json<unknown>(
-      `/code/sessions/${encodeURIComponent(sessionId)}/queue-paused`,
-      {
-        method: "PUT",
-        headers: this.headers(true),
-        body: JSON.stringify({ paused }),
-      },
-      204,
-    );
-  }
-
-  /** Release a paused queue so the worker starts the head row. */
-  async sendCodeQueuedNow(sessionId: string): Promise<void> {
-    await this.json<unknown>(
-      `/code/sessions/${encodeURIComponent(sessionId)}/queued/send-now`,
-      { method: "POST", headers: this.headers() },
-      204,
-    );
-  }
-
-  async listCodeWorkspaceTree(
-    workspaceId: string,
-    query?: { query?: string; limit?: number },
-  ): Promise<CodeWorkspaceTree> {
-    const params = new URLSearchParams();
-    if (query?.query) params.set("query", query.query);
-    if (query?.limit !== undefined) params.set("limit", String(query.limit));
-    const suffix = params.size > 0 ? `?${params}` : "";
-    return requireParsed(
-      parseCodeWorkspaceTree(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/tree${suffix}`,
-          { headers: this.headers() },
-        ),
-      ),
-      "code workspace tree",
-    );
-  }
-
-  async searchCodeWorkspace(
-    workspaceId: string,
-    query: {
-      query: string;
-      include?: string;
-      exclude?: string;
-      limit?: number;
-      history?: boolean;
-    },
-  ): Promise<CodeWorkspaceSearch> {
-    const params = new URLSearchParams({ query: query.query });
-    if (query.include) params.set("include", query.include);
-    if (query.exclude) params.set("exclude", query.exclude);
-    if (query.limit !== undefined) params.set("limit", String(query.limit));
-    if (query.history) params.set("history", "true");
-    return requireParsed(
-      parseCodeWorkspaceSearch(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/search?${params}`,
-          { headers: this.headers() },
-        ),
-      ),
-      "code workspace search",
-    );
-  }
-
-  async listCodeWorkspaceFiles(
-    workspaceId: string,
-    turnId?: string,
-  ): Promise<CodeWorkspaceFiles> {
-    const query = turnId ? `?turn=${encodeURIComponent(turnId)}` : "";
-    return requireParsed(
-      parseCodeWorkspaceFiles(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/files${query}`,
-          { headers: this.headers() },
-        ),
-      ),
-      "code workspace files",
-    );
-  }
-
-  async getCodeWorkspaceBlob(
-    workspaceId: string,
-    path: string,
-  ): Promise<CodeWorkspaceBlob> {
-    const params = new URLSearchParams({ path });
-    return requireParsed(
-      parseCodeWorkspaceBlob(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/blob?${params}`,
-          { headers: this.headers() },
-        ),
-      ),
-      "code workspace blob",
-    );
-  }
-
-  async commitCodeWorkspace(
-    workspaceId: string,
-    message?: string,
-  ): Promise<CodeCommitSnapshot> {
-    return requireParsed(
-      parseCodeCommit(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/git/commit`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify(message ? { message } : {}),
-          },
-        ),
-      ),
-      "code commit",
-    );
-  }
-
-  async pushCodeWorkspace(workspaceId: string): Promise<CodePushSnapshot> {
-    return requireParsed(
-      parseCodePush(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/git/push`,
-          { method: "POST", headers: this.headers() },
-        ),
-      ),
-      "code push",
-    );
-  }
-
-  async createCodePullRequest(
-    workspaceId: string,
-    body: { title?: string; body?: string } = {},
-  ): Promise<CodeWorkspacePrSnapshot> {
-    const snapshot = requireParsed(
-      parseCodeWorkspacePr(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/git/pr`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify(body),
-          },
-        ),
-      ),
-      "code pull request",
-    );
-    this.rememberWorkspaceMergeTarget(workspaceId, snapshot);
-    return snapshot;
-  }
-
-  async getCodeWorkspacePr(
-    workspaceId: string,
-  ): Promise<CodeWorkspacePrSnapshot> {
-    const snapshot = requireParsed(
-      parseCodeWorkspacePr(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/pr`,
-          {
-            headers: this.headers(),
-          },
-        ),
-      ),
-      "code pull request",
-    );
-    this.rememberWorkspaceMergeTarget(workspaceId, snapshot);
-    return snapshot;
-  }
-
-  /** Every pull request attributed to the workspace (decision 77). */
-  async getCodeWorkspacePullRequests(
-    workspaceId: string,
-  ): Promise<CodeWorkspacePullRequests> {
-    return requireParsed(
-      parseCodeWorkspacePullRequests(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/pull-requests`,
-          {
-            headers: this.headers(),
-          },
-        ),
-      ),
-      "code workspace pull requests",
-    );
-  }
-
-  /** Force a fresh host read, bypassing the server's short PR cache. */
-  async refreshCodeWorkspacePr(
-    workspaceId: string,
-  ): Promise<CodeWorkspacePrSnapshot> {
-    const snapshot = requireParsed(
-      parseCodeWorkspacePr(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/pr/refresh`,
-          { method: "POST", headers: this.headers() },
-        ),
-      ),
-      "code pull request",
-    );
-    this.rememberWorkspaceMergeTarget(workspaceId, snapshot);
-    return snapshot;
-  }
-
-  /**
-   * Download the failing checks' job logs into private storage.
-   *
-   * The fix-errors action calls this before it prompts, so the agent opens a
-   * file instead of working out which job failed and asking GitHub itself.
-   * Each call replaces the previous set: only the head being fixed matters.
-   */
-  async writeCodeCheckLogs(
-    workspaceId: string,
-  ): Promise<CodeCheckLogsSnapshot> {
-    return requireParsed(
-      parseCodeCheckLogsSnapshot(
-        // A GitHub read, so it takes the delivery timeout rather than hanging
-        // the button on a `gh` that never answers.
-        await this.deliveryJson(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/pr/check-logs`,
-          { method: "POST", headers: this.headers() },
-        ),
-      ),
-      "code check logs",
-    );
-  }
-
-  /** Triggers armed on a repository, enabled or not. */
-  async listCodeTriggers(repoId: string): Promise<CodeTriggerSnapshot[]> {
-    return requireParsed(
-      parseCodeTriggers(
-        await this.json(`/code/repos/${encodeURIComponent(repoId)}/triggers`, {
-          headers: this.headers(),
-        }),
-      ),
-      "code triggers",
-    );
-  }
-
-  /**
-   * Arm a trigger. Re-arming a condition updates the rule already there
-   * rather than adding a second one, so this is safe to call twice.
-   */
-  async createCodeTrigger(
-    repoId: string,
-    condition: CodeTriggerCondition,
-    action: CodeTriggerAction,
-  ): Promise<CodeTriggerSnapshot> {
-    return requireParsed(
-      parseCodeTrigger(
-        await this.json(`/code/repos/${encodeURIComponent(repoId)}/triggers`, {
-          method: "POST",
-          headers: this.headers(true),
-          body: JSON.stringify({ condition, action }),
-        }),
-      ),
-      "code trigger",
-    );
-  }
-
-  /** Switch a trigger on or off. The rule survives either way. */
-  async setCodeTriggerEnabled(
-    repoId: string,
-    triggerId: string,
-    enabled: boolean,
-  ): Promise<CodeTriggerSnapshot> {
-    return requireParsed(
-      parseCodeTrigger(
-        await this.json(
-          `/code/repos/${encodeURIComponent(repoId)}/triggers/${encodeURIComponent(triggerId)}`,
-          {
-            method: "PATCH",
-            headers: this.headers(true),
-            body: JSON.stringify({ enabled }),
-          },
-        ),
-      ),
-      "code trigger",
-    );
-  }
-
-  /** Remove a trigger. Its recorded fires stay in the transcript. */
-  deleteCodeTrigger(repoId: string, triggerId: string): Promise<void> {
-    return this.json(
-      `/code/repos/${encodeURIComponent(repoId)}/triggers/${encodeURIComponent(triggerId)}`,
-      { method: "DELETE", headers: this.headers() },
-    );
-  }
-
-  /** Start a durable watch on the workspace's open pull request. */
-  async startCodeWatch(workspaceId: string): Promise<CodeWatchSnapshot> {
-    return requireParsed(
-      parseCodeWatch(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/watch`,
-          { method: "POST", headers: this.headers() },
-        ),
-      ),
-      "code watch",
-    );
-  }
-
-  /** Stop the workspace's active watch and end its session. */
-  async stopCodeWatch(workspaceId: string): Promise<CodeWatchSnapshot> {
-    return requireParsed(
-      parseCodeWatch(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/watch`,
-          { method: "DELETE", headers: this.headers() },
-        ),
-      ),
-      "code watch",
-    );
-  }
-
-  async getCodePrComments(
-    workspaceId: string,
-  ): Promise<CodePrCommentsSnapshot> {
-    return requireParsed(
-      parseCodePrComments(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/pr/comments`,
-          { headers: this.headers() },
-        ),
-      ),
-      "code pull request comments",
-    );
-  }
-
-  /**
-   * User-initiated draft-to-ready. Decision 42 keeps pull-request state
-   * changes off the agent path, so this is a dedicated endpoint rather than a
-   * prompt. Returns the post-change snapshot.
-   */
-  async markCodePrReady(workspaceId: string): Promise<CodeWorkspacePrSnapshot> {
-    const snapshot = requireParsed(
-      parseCodeWorkspacePr(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/pr/ready`,
-          { method: "POST", headers: this.headers(true) },
-        ),
-      ),
-      "code pull request",
-    );
-    this.rememberWorkspaceMergeTarget(workspaceId, snapshot);
-    return snapshot;
-  }
-
-  /**
-   * User-initiated merge. auto=true arms host auto-merge instead of merging
-   * immediately. Returns the post-merge snapshot.
-   */
-  async mergeCodePr(
-    workspaceId: string,
-    body:
-      | CodeWorkspaceMergeRequest
-      | { method: CodePrMergeMethod; auto?: boolean },
-  ): Promise<CodeWorkspacePrSnapshot> {
-    const exact =
-      "target" in body && "expected_head_sha" in body
-        ? {
-            target: body.target,
-            expected_head_sha: body.expected_head_sha,
-          }
-        : this.workspaceMergeTargets.get(workspaceId);
-    if (!exact) {
-      throw new HttpError(
-        409,
-        "409: Refresh the pull request before merging it.",
-        "pr_identity_missing",
-      );
-    }
-    const request: CodeWorkspaceMergeRequest = {
-      ...exact,
-      method: body.method,
-      auto: body.auto ?? false,
-    };
-    const result = parseCodeWorkspaceMergeResult(
-      await this.json(
-        `/code/workspaces/${encodeURIComponent(workspaceId)}/pr/merge`,
-        {
-          method: "POST",
-          headers: this.headers(true),
-          body: JSON.stringify(request),
-        },
-      ),
-      request,
-    );
-    this.rememberWorkspaceMergeTarget(workspaceId, result.status);
-    return result.status;
-  }
-
-  private rememberWorkspaceMergeTarget(
-    workspaceId: string,
-    snapshot: CodeWorkspacePrSnapshot,
-  ): void {
-    const exact = snapshot.pr ? mergeTargetFromDigest(snapshot.pr) : null;
-    if (exact) this.workspaceMergeTargets.set(workspaceId, exact);
-    else this.workspaceMergeTargets.delete(workspaceId);
-  }
-
-  async runCodeWorkspaceAction(
-    workspaceId: string,
-    name: string,
-  ): Promise<CodeActionSnapshot> {
-    return requireParsed(
-      parseCodeAction(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/actions/${encodeURIComponent(name)}`,
-          { method: "POST", headers: this.headers() },
-        ),
-      ),
-      "code quick action",
-    );
-  }
-
-  async getCodeWorkspaceDiff(
-    workspaceId: string,
-    opts: { turn?: string; file?: string } = {},
-  ): Promise<CodeWorkspaceDiff> {
-    const params = new URLSearchParams();
-    if (opts.turn) params.set("turn", opts.turn);
-    if (opts.file) params.set("file", opts.file);
-    const query = params.size > 0 ? `?${params.toString()}` : "";
-    return requireParsed(
-      parseCodeWorkspaceDiff(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/diff${query}`,
-          { headers: this.headers() },
-        ),
-      ),
-      "code workspace diff",
-    );
-  }
-
-  /**
-   * Write one fork of a session — the condensed transcript plus per-turn
-   * records — into private storage, for a child agent to read.
-   *
-   * `atTurnId` forks at the end of that turn; omitted, the fork covers the
-   * whole conversation. The child session is created separately: this call
-   * only produces the files, so the reader still picks the engine and edits
-   * the framing before anything is sent.
-   */
-  async forkCodeSession(
-    sessionId: string,
-    atTurnId?: string,
-  ): Promise<CodeForkTranscript> {
-    return requireParsed(
-      parseCodeForkTranscript(
-        await this.json(
-          `/code/sessions/${encodeURIComponent(sessionId)}/fork`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify(atTurnId ? { at_turn: atTurnId } : {}),
-          },
-        ),
-      ),
-      "code fork transcript",
-    );
-  }
-
-  async reapCodeSession(sessionId: string): Promise<CodeSessionSnapshot> {
-    return requireParsed(
-      parseCodeSession(
-        await this.json(
-          `/code/sessions/${encodeURIComponent(sessionId)}/reap`,
-          {
-            method: "POST",
-            headers: this.headers(),
-          },
-        ),
-      ),
-      "code session",
-    );
-  }
-
-  /** Every adapter grant the owner holds, revoked rows included. */
-  async listCodeGrants(): Promise<CodeGrantSnapshot[]> {
-    return requireParsed(
-      parseCodeGrantList(
-        await this.json("/code/grants", { headers: this.headers() }),
-      ),
-      "code grants",
-    );
-  }
-
-  async revokeCodeGrant(
-    grantId: string,
-    reason?: string,
-  ): Promise<CodeGrantSnapshot> {
-    return requireParsed(
-      parseCodeGrant(
-        await this.json(`/code/grants/${encodeURIComponent(grantId)}/revoke`, {
-          method: "POST",
-          headers: this.headers(true),
-          body: JSON.stringify(reason ? { reason } : {}),
-        }),
-      ),
-      "code grant",
-    );
-  }
-
-  /** Revoke every live grant one channel workspace holds. */
-  async revokeCodeGrantWorkspace(
-    channelKind: string,
-    workspaceIdentity: string,
-  ): Promise<CodeGrantSnapshot[]> {
-    return requireParsed(
-      parseCodeGrantList(
-        await this.json("/code/grants/revoke-workspace", {
-          method: "POST",
-          headers: this.headers(true),
-          body: JSON.stringify({
-            channel_kind: channelKind,
-            workspace_identity: workspaceIdentity,
-          }),
-        }),
-      ),
-      "code grants",
-    );
-  }
-
-  /** What the connect approval page renders; a used or stale link 404s. */
-  async getCodeConnectPage(nonce: string): Promise<CodeConnectPage> {
-    return requireParsed(
-      parseCodeConnectPage(
-        await this.json(`/external/connect/${encodeURIComponent(nonce)}`, {
-          headers: this.headers(),
-        }),
-      ),
-      "connect page",
-    );
-  }
-
-  /** The owner's "is this you?". Mints nothing by itself. */
-  approveCodeConnect(nonce: string, csrf: string): Promise<void> {
-    return this.json(`/external/connect/${encodeURIComponent(nonce)}/approve`, {
-      method: "POST",
-      headers: this.headers(true),
-      body: JSON.stringify({ csrf }),
-    });
-  }
-
-  async listCodeTerminals(
-    workspaceId: string,
-  ): Promise<CodeTerminalSnapshot[]> {
-    const body = await this.json<unknown>(
-      `/code/workspaces/${encodeURIComponent(workspaceId)}/terminals`,
-      { headers: this.headers() },
-    );
-    return requireParsed(parseCodeTerminalList(body), "code terminals");
-  }
-
-  async createCodeTerminal(
-    workspaceId: string,
-    body: { cols?: number; rows?: number } = {},
-  ): Promise<CodeTerminalSnapshot> {
-    return requireParsed(
-      parseCodeTerminal(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/terminals`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify(body),
-          },
-        ),
-      ),
-      "code terminal",
-    );
-  }
-
-  deleteCodeTerminal(workspaceId: string, terminalId: string): Promise<void> {
-    return this.json(
-      `/code/workspaces/${encodeURIComponent(workspaceId)}/terminals/${encodeURIComponent(terminalId)}`,
-      { method: "DELETE", headers: this.headers() },
-    );
-  }
-
-  async readCodeTerminal(
-    workspaceId: string,
-    terminalId: string,
-    cursor = 0,
-  ): Promise<CodeTerminalRead> {
-    return requireParsed(
-      parseCodeTerminalRead(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/terminals/${encodeURIComponent(terminalId)}/read?cursor=${encodeURIComponent(String(cursor))}`,
-          { headers: this.headers() },
-        ),
-      ),
-      "code terminal read",
-    );
-  }
-
-  writeCodeTerminal(
-    workspaceId: string,
-    terminalId: string,
-    data: string,
-  ): Promise<void> {
-    return this.json(
-      `/code/workspaces/${encodeURIComponent(workspaceId)}/terminals/${encodeURIComponent(terminalId)}/write`,
-      {
-        method: "POST",
-        headers: this.headers(true),
-        body: JSON.stringify({ bytes: encodeUtf8Base64(data) }),
-      },
-    );
-  }
-
-  async resizeCodeTerminal(
-    workspaceId: string,
-    terminalId: string,
-    cols: number,
-    rows: number,
-  ): Promise<CodeTerminalSnapshot> {
-    return requireParsed(
-      parseCodeTerminal(
-        await this.json(
-          `/code/workspaces/${encodeURIComponent(workspaceId)}/terminals/${encodeURIComponent(terminalId)}/resize`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify({ cols, rows }),
-          },
-        ),
-      ),
-      "code terminal",
-    );
-  }
-
-  /** Open the per-session journal; auth via Sec-WebSocket-Protocol. */
-  openCodeEvents(
-    sessionId: string,
-    after: number,
-    onFrame: (frame: SequencedCodeEventFrame) => void,
-  ): WebSocket {
-    const url = `${this.baseUrl.replace(/^http/, "ws")}/code/sessions/${encodeURIComponent(sessionId)}/events?after=${after}`;
-    const protocols = [WS_HANDSHAKE, `${WS_TOKEN_PREFIX}${this.token}`];
-    const socket = new WebSocket(url, protocols);
-    socket.onmessage = (msg) => {
-      try {
-        const frame = parseSequencedCodeEvent(JSON.parse(String(msg.data)));
-        if (frame) onFrame(frame);
-        else console.error("dropping malformed code event frame");
-      } catch (err) {
-        console.error("bad code event frame", err);
-      }
-    };
-    return socket;
-  }
-
-  /** Open the install-wide digest channel; auth via Sec-WebSocket-Protocol. */
-  openCodeUpdates(onNotice: (notice: CodeUpdateNotice) => void): WebSocket {
-    const url = `${this.baseUrl.replace(/^http/, "ws")}/code/updates`;
-    const protocols = [WS_HANDSHAKE, `${WS_TOKEN_PREFIX}${this.token}`];
-    const socket = new WebSocket(url, protocols);
-    socket.onmessage = (msg) => {
-      try {
-        const notice = parseCodeUpdateNotice(JSON.parse(String(msg.data)));
-        if (notice) onNotice(notice);
-        else console.error("dropping malformed code update notice");
-      } catch (err) {
-        console.error("bad code update notice", err);
-      }
-    };
-    return socket;
-  }
-
-  async setCodeAttention(
-    sessionId: string,
-    body: { clear?: boolean; note?: string },
-  ): Promise<CodeSessionSnapshot> {
-    return requireParsed(
-      parseCodeSession(
-        await this.json(
-          `/code/sessions/${encodeURIComponent(sessionId)}/attention`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify(body),
-          },
-        ),
-      ),
-      "code attention",
-    );
-  }
-
-  async listCodeApprovals(query?: {
-    state?: "pending" | "approved" | "denied";
-    sessionId?: string;
-  }): Promise<CodeApprovalSnapshot[]> {
-    const params = new URLSearchParams();
-    if (query?.state) params.set("state", query.state);
-    if (query?.sessionId) params.set("session_id", query.sessionId);
-    const suffix = params.size > 0 ? `?${params}` : "";
-    const body = await this.json<unknown>(`/code/approvals${suffix}`, {
-      headers: this.headers(),
-    });
-    return parseList(body, parseCodeApproval, "code approvals");
-  }
-
-  async decideCodeApproval(
-    approvalId: string,
-    body: { decision: "approve" | "deny"; feedback?: string },
-  ): Promise<CodeApprovalSnapshot> {
-    return requireParsed(
-      parseCodeApproval(
-        await this.json(
-          `/code/approvals/${encodeURIComponent(approvalId)}/decision`,
-          {
-            method: "POST",
-            headers: this.headers(true),
-            body: JSON.stringify(body),
-          },
-        ),
-      ),
-      "code approval",
-    );
-  }
-}
-
-function requireParsed<T>(value: T | null, label: string): T {
-  if (!value) throw new Error(`${label} response contains invalid data`);
-  return value;
-}
-
-function mergeTargetFromDigest(
-  pr: PullRequestDigest,
-): Pick<CodeWorkspaceMergeRequest, "target" | "expected_head_sha"> | null {
-  if (!pr.url || !pr.head_sha) return null;
-  let url: URL;
-  try {
-    url = new URL(pr.url);
-  } catch {
-    return null;
-  }
-  const parts = url.pathname.split("/").filter(Boolean);
-  if (
-    parts.length !== 4 ||
-    parts[2] !== "pull" ||
-    Number(parts[3]) !== pr.number ||
-    !url.hostname
-  ) {
-    return null;
-  }
-  return {
-    target: {
-      repository: {
-        host: url.hostname,
-        owner: parts[0],
-        name: parts[1],
-      },
-      number: pr.number,
-    },
-    expected_head_sha: pr.head_sha,
-  };
-}
-
-function parseCodeWorkspaceMergeResult(
-  value: unknown,
-  expected: CodeWorkspaceMergeRequest,
-): CodeWorkspaceMergeResult {
-  if (typeof value !== "object" || value === null || Array.isArray(value)) {
-    throw new Error("code pull request merge response contains invalid data");
-  }
-  const record = value as Record<string, unknown>;
-  const target = record.target as
-    | {
-        repository?: { host?: unknown; owner?: unknown; name?: unknown };
-        number?: unknown;
-      }
-    | undefined;
-  const repository = target?.repository;
-  const host = repository?.host;
-  const owner = repository?.owner;
-  const name = repository?.name;
-  const acceptedHead = record.accepted_head_sha;
-  const status = parseCodeWorkspacePr(record.status);
-  if (
-    !repository ||
-    typeof host !== "string" ||
-    typeof owner !== "string" ||
-    typeof name !== "string" ||
-    typeof target.number !== "number" ||
-    typeof acceptedHead !== "string" ||
-    !status ||
-    !sameRepository({ host, owner, name }, expected.target.repository) ||
-    target.number !== expected.target.number ||
-    acceptedHead !== expected.expected_head_sha
-  ) {
-    throw new Error("code pull request merge response contains invalid data");
-  }
-  return {
-    target: {
-      repository: {
-        host,
-        owner,
-        name,
-      },
-      number: target.number,
-    },
-    accepted_head_sha: acceptedHead,
-    status,
-  };
-}
-
-function sameRepository(
-  left: { host: string; owner: string; name: string },
-  right: { host: string; owner: string; name: string },
-): boolean {
-  return (
-    left.host.toLowerCase() === right.host.toLowerCase() &&
-    left.owner.toLowerCase() === right.owner.toLowerCase() &&
-    left.name.toLowerCase() === right.name.toLowerCase()
-  );
-}
-
-function encodeUtf8Base64(text: string): string {
-  const bytes = new TextEncoder().encode(text);
-  let binary = "";
-  for (const byte of bytes) binary += String.fromCharCode(byte);
-  return btoa(binary);
-}
-
-function parseList<T>(
-  body: unknown,
-  parse: (value: unknown) => T | null,
-  label: string,
-): T[] {
-  if (!Array.isArray(body)) {
-    throw new Error(`${label} response is not an array`);
-  }
-  return body.map((item, index) => {
-    const parsed = parse(item);
-    if (!parsed) {
-      throw new Error(`${label} response contains invalid data at ${index}`);
-    }
-    return parsed;
-  });
 }
